### 2024-04-02 #6576 @editorialbot
[PRE REVIEW]: OpenCCM: An Open-Source Continuous Compartmental Modelling Package
**Submitting author:** @nasserma (<a href="http://orcid.org/0000-0002-1772-0376">Nasser Mohieddin Abukhdeir</a>)
**Repository:** https://github.com/uw-comphys/openccm
**Branch with paper.md** (empty if default branch):
**Version:** v0.9
