### 2024-04-02 #6576 @danielskatz
👋 @nasserma - thanks for your submission. Note that your paper is about 2.5x the maximum recommended length for JOSS papers.  Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). I'll put this on pause for now. Let me know when it's ready to restart.

In addition, please fix the DOIs with additional prefixes in your bib file. After you make changes to your .bib file, use the command `@editorialbot check references` to check them again. And use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
