### 2024-04-02 98f24ac
- corrected DOIs in JOSS submission bib file - significantly reduced the length of the JOSS manuscript to meet length restrictions.
