+++ publications/2403_JOSS/paper.bib
@@ -50,7 +50,7 @@ volume = {185},
 pages = {108650},
 year = {2024},
 issn = {0098-1354},
-doi = {https://doi.org/10.1016/j.compchemeng.2024.108650},
+doi = {10.1016/j.compchemeng.2024.108650},
 url = {https://www.sciencedirect.com/science/article/pii/S0098135424000681},
 author = {Alexandru Andrei Vasile and Marc G. Aucoin and Hector Budman and Nasser Mohieddin Abukhdeir},
 keywords = {Compartmental modelling, Reduced order modelling, Fluid dynamics, Systemic model},
@@ -63,7 +63,7 @@ volume = {210},
 pages = {115196},
 year = {2019},
 issn = {0009-2509},
-doi = {https://doi.org/10.1016/j.ces.2019.115196},
+doi = {10.1016/j.ces.2019.115196},
 url = {https://www.sciencedirect.com/science/article/pii/S0009250919306888},
 author = {Nicolas Jourdan and Thibaut Neveux and Olivier Potier and Mohamed Kanniche and Jim Wicks and Ingmar Nopens and Usman Rehman and Yann {Le Moullec}},
 keywords = {Compartmental Modelling, Hydrodynamics, Multi-scale, CFD, Systemic model},
@@ -78,7 +78,7 @@ keywords = {Compartmental Modelling, Hydrodynamics, Multi-scale, CFD, Systemic m
   isbn      = {9781119176817},
   abstract  = {Abstract This chapter presents an overview of model order reduction – a new paradigm in the field of simulation-based engineering sciences, and one that can tackle the challenges and leverage the opportunities of modern ICT technologies. Despite the impressive progress attained by simulation capabilities and techniques, a number of challenging problems remain intractable. These problems are of different nature, but are common to many branches of science and engineering. Among them are those related to high-dimensional problems, problems involving very different time scales, models defined in degenerate domains with at least one of the characteristic dimensions much smaller than the others, model requiring real-time simulation, and parametric models. All these problems represent a challenge for standard mesh-based discretization techniques; yet the ability to solve these problems efficiently would open unexplored routes for real-time simulation, inverse analysis, uncertainty quantification and propagation, real-time optimization, and simulation-based control – critical needs in many branches of science and engineering. Model order reduction offers new simulation alternatives by circumventing, or at least alleviating, otherwise intractable computational challenges. In the present chapter, we revisit three of these model reduction techniques: proper orthogonal decomposition, proper generalized decomposition, and reduced basis methodologies.},
   booktitle = {Encyclopedia of Computational Mechanics Second Edition},
-  doi       = {https://doi.org/10.1002/9781119176817.ecm2110},
+  doi       = {10.1002/9781119176817.ecm2110},
   eprint    = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781119176817.ecm2110},
   keywords  = {model order reduction, parametric models, non-incremental solvers, Real time simulation, propoer orthogonal decomposition, reduced basis, proper generalized decomposition, separated representations, greedy algorithms},
   url       = {https://onlinelibrary.wiley.com/doi/abs/10.1002/9781119176817.ecm2110},
