### 2024-04-02 #6576 @danielskatz
You did this correctly in https://github.com/openjournals/joss-reviews/issues/6576#issuecomment-2032995975 and just above as well. This track editor should now be aware of this and will get back to you when they have a chance.
