### 2024-04-08 e5286a7
Significantly speed up _calculate_facet_normals and _calculate_facet_sizes
