### 2024-04-09 ab25f0b
Change facet_to_bc_map to store ints (as it should have been doing)
