### 2024-04-10 e0aae7f
Added support for use of alphanumeric specie names
