+++ openccm/system_solvers/reactions.py
@@ -16,6 +16,7 @@
 ########################################################################################################################
 
 import inspect
+from string import ascii_lowercase as alphabet
 from typing import List, Optional, Tuple
 
 import numpy as np
@@ -100,9 +101,6 @@ def parse_reactions(specie_order: List[str], all_reactions: List[str], all_rates
 
     See the Notes section below for further description of the expectations and limitations of this custom reactions configuration parser.
 
-    Only forward WRITTEN reactions are supported (use of ->). I.e. reversible reactions should be written as 2 separate forward reactions. Thus use of <-> is not supported.
-    Molecules with standard chemical superscripts (i.e. ions) and subscripts (i.e. compounds) are not supported by this parser. I.e., if you want to write 5O2 (i.e. 5 dioxygen molecules), you can use 5A instead, where A (simulation) = O2 (reality).
-
     Args:
         specie_order:       The ordered species as written in the main configuration file. Must match the same types of species present in reactions config file.
         all_reactions:      The partially-parsed reactions as extracted from reactions configuration file.
@@ -113,9 +111,10 @@ def parse_reactions(specie_order: List[str], all_reactions: List[str], all_rates
     
     Notes:
         The reaction parser only supports forward (written) reactions (i.e. use of ->). Reversible reactions must be written as two independent reactions with their associated rate constants (see example above).
-        Furthermore, all chemical species must be written as letter characters (and not alphanumeric). I.e., standard chemical species with superscripts (ions) or subscripts (compounds) are not supported.
-        For example, if a reaction contains 5O2 (5 dioxygen molecules), it must be written as 5A or 5OTWO in the reactions config (i.e. no trailing numbers or underscores).
+        Furthermore, all chemical species must be strictly alphanumeric and must not contain any special characters (i.e. _, -, +).
+        I.e., if wishing to use the species (in traditional chemistry notation) Na_2CO_3, this must be written as Na2CO3.
     """ 
+
     # Parse the reaction ids (i.e. R1, R2 etc) from the actual reactions and associated rates. These are needed for proper input checking.
     rate_ids = [rate.split(':')[0].strip() for rate in all_rates]
     rxn_ids = [rxn.split(':')[0].strip() for rxn in all_reactions]
@@ -160,6 +159,20 @@ def parse_reactions(specie_order: List[str], all_reactions: List[str], all_rates
             raise ValueError(f"{e}. Please check that rate constants are numeric in reactions config.")
         # If rate constant is numeric, then append to reaction book
         rxn_book[curr_key].append(rate_book.get(curr_key))
+
+    # Create mapping of species to dummy specie names which is much easier for parsing process
+    specie_order_dummy = list(alphabet[0:len(specie_order)])
+    dummy_map = dict(zip(specie_order, specie_order_dummy))
+    rxn_book_keys = list(rxn_book.keys())
+    for key in rxn_book_keys:
+        # get current reaction
+        rxn = rxn_book[key][0]
+        for spec in specie_order:
+            # replace species with dummy specie names
+            rxn = rxn.replace(spec, dummy_map[spec])
+        
+        # assign transformed reaction back to master reaction book
+        rxn_book[key][0] = rxn
 
     # The following pyparsing structure outlines the expected reaction structure from reactions config file.
     rxnType = pp.ZeroOrMore(pp.Literal('->')) 
@@ -233,6 +246,18 @@ def parse_reactions(specie_order: List[str], all_reactions: List[str], all_rates
     # Until this point, the LHS was written as dA (i.e. for dA/dt). shorten this notation further such that dA --> A
     DEsysLHS = [LHS[1:] for LHS in tempLHS]
     DEsysRHS = tempRHS.tolist()
+
[…]
