### 2024-04-10 af213a8
Cleanup (#20)
* Stop storing average direction, calculate it once right before we need it.

* Cleanup of `connect_compartments_using_unidirectional_assumptions`
