### 2024-04-10 dce6a91
Split parser function into two for readability
