### 2024-04-11 b0de2c2
Added test which runs all of the examples to ensure they pass their asserts.
