### 2024-04-11 68f4e0f
Added warning to compartmentalization and merging steps if too many compartments are created or merged as this likely represents misspecified tolerances.
