### 2024-04-12 fcf39e8
Stop modifying CMesh as part of the compartmentalization.
