### 2024-04-12 be5f5b3
Switch element_connectivity to immutable since it's no longer changed.
