### 2024-04-12 7bb3b4d
Fixed a bug with how the average compartment direction was calculated
