### 2024-04-15 4930dee
Restructure merging of compartments.
