### 2024-04-15 75c6109
Changing how merging of compartments functions in order to prevent cstr compartments with all inlets or all outlets from being created.
