+++ openccm/compartmentalize/unidirectional.py
@@ -78,9 +78,8 @@ def create_compartment_network(compartments:    Dict[int, Set[int]],
             element_to_compartment_map[element] = id_compartment
 
     # Add BCs to the element compartment_map
-    for _id in mesh.grouped_bcs.domain_inlets + mesh.grouped_bcs.domain_outlets + mesh.grouped_bcs.ignored:
+    for _id in mesh.grouped_bcs.domain_inlets + mesh.grouped_bcs.domain_outlets + mesh.grouped_bcs.ignored + (mesh.grouped_bcs.no_flux,):
         element_to_compartment_map[_id] = _id
-    element_to_compartment_map[mesh.grouped_bcs.no_flux] = mesh.grouped_bcs.no_flux
 
     with open(log_folder_path + "network.txt", 'w') as logging:
         # Group the bounding entities of this compartment based on the compartment on the other side of them
@@ -99,7 +98,7 @@ def create_compartment_network(compartments:    Dict[int, Set[int]],
                 # If the facet is on the bounds of the mesh, then the element_id will be negative
                 element_on_other_side_of_bound = bounding_facets_info[facet][1]
 
-                if element_on_other_side_of_bound == mesh.grouped_bcs.no_flux or \
+                if element_on_other_side_of_bound == mesh.grouped_bcs.no_flux or element_on_other_side_of_bound in mesh.grouped_bcs.ignored or \
                         element_on_other_side_of_bound not in element_to_compartment_map:
                     bounding_facets_info.pop(facet)
 
@@ -350,8 +349,8 @@ def merge_compartments(compartments:        Dict[int, Set[int]],
         assert len(network) > 0
         assert compartment_id not in network
 
-    DEBUG                = config_parser.get_item(['SETUP',                'DEBUG'],                bool)
-    log_folder_path      = config_parser.get_item(['SETUP',                'log_folder_path'],      str)
+    DEBUG                = config_parser.get_item(['SETUP',                 'DEBUG'],                bool)
+    log_folder_path      = config_parser.get_item(['SETUP',                 'log_folder_path'],      str)
     # The minimum size of a compartment. Units are the same as those used by the mesh.
     min_compartment_size = config_parser.get_item(['COMPARTMENTALIZATION',  'min_compartment_size'], float)
     model                = config_parser.get_item(['COMPARTMENT MODELLING', 'model'],                str)
@@ -387,16 +386,6 @@ def merge_compartments(compartments:        Dict[int, Set[int]],
     else:
         raise ValueError(f"Unsupported model type: {model}")
 
-    # Remove all connections that lead to a BC since we do not want to consider them for the purpose of merging compartments
-    for _id_pfr_tmp in connection_pairing:
-        _connections_for_pfr = connection_pairing[_id_pfr_tmp]
-        for i_connection in list(_connections_for_pfr.keys()):  # Need to make the copy with list() since we're modifying the dict during the loop
-            if _connections_for_pfr[i_connection] < 0:  # Negative PFR if represents a
-                _connections_for_pfr.pop(i_connection)
-        # If there are no connections available after removing the domain boundaries, then this compartment cannot be merged with something else
-        if len(_connections_for_pfr) == 0:
-            # Setting to infinite so that a numpy array indexed by compartment id can be used
-            compartment_sizes[_id_pfr_tmp] = np.inf
     with open(log_folder_path + "merge.txt", 'w') as logging:
         if DEBUG:
             logging.write("[ Pre\n")
@@ -405,70 +394,48 @@ def merge_compartments(compartments:        Dict[int, Set[int]],
             logging.write("]\n")
             logging.write("[ Merging\n")
 
-        # Iterate over compartments, and merge together any that only have one connection.
+        # Merge all compartments which have only one neighbour or all inlets/outlet.
[…]
