### 2024-04-16 e7d814b
Skip attempting to merge compartments only connected to domain boundaries (e.g. a compartment running from a domain inlet to a domain outlet)
