### 2024-04-16 e48cb01
Add immutable flags to CMesh's numpy objects
