### 2024-04-16 0607830
Changed compartments_to_pfrs' signature to not require compartment_network, making it easier to reason about where data is being used.
