### 2024-04-16 69b344f
Rewrite flow optimization to not require rtol_opt.
