+++ openccm/compartment_models/helpers.py
@@ -26,14 +26,13 @@ def tweak_compartment_flows(
         connection_pairing: Dict[int, Dict[int, int]],
         volumetric_flows: Dict[int, float],
         grouped_bcs: GroupedBCs,
-        atol_opt: float,
-        rtol_opt: float
+        atol_opt: float
 ) -> None:
     """
     This function is used to adjust the flowrates in the compartment network so that the net flow around a compartment
     is either 0 or slightly positive.
 
-    This is different from `tweak_cstr_pfr_flows` in a few major ways:
+    This is different from `tweak_final_flows` in a few major ways:
     - This method works balances the mass around each COMPARTMENT, that one balances the mass around each CSTR/PFR.
     - That method tries to minimize the error in the conservation of mass, accepting small deviations resulting in either
       a small net inflow or outflow.
@@ -79,8 +78,7 @@ def tweak_compartment_flows(
                                     Connection ID in this dictionary is ALWAYS positive, need to take absolute sign of
                                     the value if it's negative (see `connection_pairing` docstring)
         grouped_bcs:            GroupedBCs object for identifying which connections belong to domain inlets/outlets.
-        atol_opt:               Absolute tolerance for evaluating conservation of mass of the optimized system
-        rtol_opt:               Relative tolerance for evaluating conservation of mass of the optimized system
+        atol_opt:               Absolute tolerance for evaluating conservation of mass of the optimized system.
 
     Returns:
         Nothing. Values are changed implicitly
@@ -100,7 +98,7 @@ def tweak_compartment_flows(
     # Volumetric flows cannot be assumed to be put into the dictionary in increasing order
     con_to_index = {_id: index for index, _id in enumerate(list(volumetric_flows.keys()))}
 
-    # NOTE: There are no equality constriants
+    # NOTE: There are no equality constraints
 
     # Create the inequality constraint
     A = np.zeros((len(connection_pairing), len(volumetric_flows)), dtype='b')
@@ -159,8 +157,6 @@ def tweak_compartment_flows(
 
     # Check the conservation of mass for the system as a whole and each compartment.
     # Each compartment is checked again to try to catch any issues with the optimization setup above
-    total_inflow  = 0.0
-    total_outflow = 0.0
     missing_flow = 0.0
     for compartment, compartment_connections in connection_pairing.items():
         net_flow = 0.0
@@ -168,31 +164,24 @@ def tweak_compartment_flows(
             if connection > 0:
                 net_flow += volumetric_flows[connection]
                 if compartment_other < 0:
-                    if compartment_other in grouped_bcs.domain_inlets:
-                        total_inflow += volumetric_flows[connection]
-                    else:
-                        # Subtracting on purpose
-                        missing_flow -= volumetric_flows[connection]
+                    if compartment_other not in grouped_bcs.domain_inlets:
+                        missing_flow += volumetric_flows[connection]
             else:
                 connection = abs(connection)
                 net_flow -= volumetric_flows[connection]
                 if compartment_other < 0:
-                    if compartment_other in grouped_bcs.domain_outlets:
-                        total_outflow += volumetric_flows[connection]
-                    else:
-                        missing_flow += volumetric_flows[connection]
-
+                    if compartment_other not in grouped_bcs.domain_outlets:
+                        missing_flow -= volumetric_flows[connection]
         assert net_flow > 0
-        assert np.isclose(net_flow, 0, rtol=rtol_opt, atol=atol_opt)
-    assert np.isclose(0.0, missing_flow, rtol=rtol_opt, atol=atol_opt)
[…]
