### 2024-04-16 f6521a7
Change `tweak_compartment_flows` to correctly impose atol_opt on the optimization
