### 2024-04-18 8c63ac7
FIX: Check to avoid warning about invalid values in division. There can be infinity, but they are expected because the compartment numbering can be non-sequential.
