### 2024-04-18 2b67739
NEW: Added check for connect_cstr_compartments to disable asserts when running during merge_compartments
