### 2024-04-18 05ffe0d
FIX: Correctly export compartment labels pre- and post- merging on subsequent runs.
