### 2024-04-18 5dfb7cf
FIX: Merging two compartments sometimes grabbed wrong ID
