### 2024-04-19 3513611
FIX: Check against flow threshold in connect_cstr_compartments only on final pass
