+++ openccm/compartment_models/pfr.py
@@ -112,13 +112,13 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
         connections.sort(key=lambda element: element[0])
 
         # 1.3a Reorder connections
-        connections = _fix_connection_ordering(connections, _volumetric_flows, id_compartment)
+        connections = _fix_connection_ordering(connections, _volumetric_flows, id_compartment, atol_opt)
 
         # 1.3b Re-order domain inlets/outlets
         connections = _fix_domain_boundary_connection_ordering(connections, compartment_connections, mesh.grouped_bcs, id_compartment)
 
         # 1.4 Merge connection locations
-        connections_merged = _merge_connections(connections, _volumetric_flows, dist_threshold, compartment_connections)
+        connections_merged = _merge_connections(connections, _volumetric_flows, dist_threshold, compartment_connections, atol_opt)
 
         # Save the information about the connection locations
         connection_locations[id_compartment] = connections_merged
@@ -421,8 +421,11 @@ def compartments_to_pfrs(connection_locations:      Dict[int, List[Tuple[float, 
     return volumes_pfr, volumetric_flows, all_connection_pairing, compartment_to_pfr_map
 
 
-def _merge_connections(inlets_and_outlets: List[Tuple[float, int]], volumetric_flows: Dict[int, float],
-                       dist_threshold: float, compartment_connections: Dict[int, int])\
+def _merge_connections(inlets_and_outlets:      List[Tuple[float, int]],
+                       volumetric_flows:        Dict[int, float],
+                       dist_threshold:          float,
+                       compartment_connections: Dict[int, int],
+                       atol_opt:                float)\
         -> List[Tuple[float, List[int]]]:
     """
     Function to merge inlets/outlet by changing their location along the compartment.
@@ -446,18 +449,19 @@ def _merge_connections(inlets_and_outlets: List[Tuple[float, int]], volumetric_f
     2. Weighted average (using volumetric flowrate) of all locations to merge.
 
     Args:
-        inlets_and_outlets: List of tuples containing the pre-merging inlets/outlets and their position.
-                            Each tuple has two values, first is a float and the second is an integer.
-                            <p>
-                            - The float represents the position along the length of the compartment of this inlet/outlet.
-                              From 0.0 to 1.0 (inclusive of both ends).
-                            <p>
-                            - The integer represents the ID of this inlet/outlet.
-                              A positive value signifies an inlet and a negative value signifies and outlet.
-        volumetric_flows:   Dictionary to look up the flowrate through each connection by its ID
-        dist_threshold:     The maximum distance between any two points in a merge connection location
+        inlets_and_outlets:         List of tuples containing the pre-merging inlets/outlets and their position.
+                                    Each tuple has two values, first is a float and the second is an integer.
+                                    <p>
+                                    - The float represents the position along the length of the compartment of this inlet/outlet.
+                                      From 0.0 to 1.0 (inclusive of both ends).
+                                    <p>
+                                    - The integer represents the ID of this inlet/outlet.
+                                      A positive value signifies an inlet and a negative value signifies and outlet.
+        volumetric_flows:           Dictionary to look up the flowrate through each connection by its ID
+        dist_threshold:             The maximum distance between any two points in a merge connection location
         compartment_connections:    Dictionary mapping the connection IDs for a compartment
                                     to the compartment on the other side.
+        atol_opt:                   Absolute tolerate used for conservation of mass.
[…]
