### 2024-04-22 cb9733f
FIX: _check_flow_requirement was being passed n_vec instead of v_vec.
