+++ openccm/compartment_models/pfr.py
@@ -27,9 +27,9 @@ from .helpers import tweak_compartment_flows, tweak_final_flows
 
 
 def create_pfr_network(compartments:        Dict[int, Set[int]],
-                       compartment_network: Dict[int, Dict[int, Dict[int, Tuple[int, np.ndarray]]]],
+                       compartment_network: Dict[int, Dict[int, Dict[int, int]]],
                        mesh:                CMesh,
-                       vel_vec:             np.ndarray,
+                       flows_and_upwind:             np.ndarray,
                        dir_vec:             np.ndarray,
                        config_parser:       ConfigParser)\
         -> Tuple[
@@ -57,7 +57,11 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
                                             - The 1st is the index of the element upwind of that boundary facet.
                                             - The 2nd is the outward facing unit normal for that boundary facet.
         mesh:                   The mesh containing the compartments.
-        vel_vec:                Numpy array of velocity vectors, row i is for element i.
+        flows_and_upwind:       2D object array indexed by facet ID.
+                                - 1st column is volumetric flowrate through facet.
+                                - 2nd column is a flag indicating which of a facet's elements are upwind of it.
+                                    - 0, and 1 represent the index into mesh.facet_elements[facet]
+                                    - -1 is used for boundary elements to represent
         dir_vec:                Numpy array of direction vectors, row i is for element i.
         config_parser:          The OpenCCM ConfigParser
 
@@ -89,7 +93,7 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
     # 1. Create connections between compartments
     ####################################################################################################################
     # 1.1 Initial connections
-    results_1 = connect_pfr_compartments(compartment_network, compartments, mesh, dir_vec, vel_vec, True, config_parser)
+    results_1 = connect_pfr_compartments(compartment_network, compartments, mesh, dir_vec, flows_and_upwind, True, config_parser)
     id_next_connection, connection_distances, connection_pairing, compartment_network, compartments, _volumetric_flows = results_1
 
     # 1.2 Optimize connections to prevent flow reversal when creating the intra-compartment flows.
@@ -773,10 +777,10 @@ def _fix_connection_ordering(inlets_and_outlets:    List[Tuple[float, int]],
     return inlets_and_outlets
 
 
-def _fix_domain_boundary_connection_ordering(inlets_and_outlets: List[Tuple[float, int]],
-                                             compartment_connections: Dict[int, int],
-                                             grouped_bcs: GroupedBCs,
-                                             id_compartment: int)\
+def _fix_domain_boundary_connection_ordering(inlets_and_outlets:        List[Tuple[float, int]],
+                                             compartment_connections:   Dict[int, int],
+                                             grouped_bcs:               GroupedBCs,
+                                             id_compartment:            int)\
         -> List[Tuple[float, int]]:
     """
     Currently, the merging of small compartments can result in compartments with multiple connections to domain inlets/outlets.
@@ -867,17 +871,17 @@ def _fix_domain_boundary_connection_ordering(inlets_and_outlets: List[Tuple[floa
     return inlets_and_outlets
 
 
-def connect_pfr_compartments(compartment_network:  Dict[int, Dict[int, Dict[int, Tuple[int, np.ndarray]]]],
-                             compartments:         Dict[int, Set[int]],
-                             mesh:                 CMesh,
-                             dir_vec:              np.ndarray,
-                             vel_vec:              np.ndarray,
-                             final:                bool,
-                             config_parser:        ConfigParser) \
+def connect_pfr_compartments(compartment_network:   Dict[int, Dict[int, Dict[int, int]]],
+                             compartments:          Dict[int, Set[int]],
[…]

[…]
