### 2024-04-22 25b40e9
FIX: Correctly merge in connection_pairing when a compartment has a connection to a BC.
