### 2024-04-24 4521c55
FIX: Correctly identify and merge compartments with only one neighbour
