### 2024-04-24 fadffcf
FORMAT: Minor readability cleanup
