### 2024-04-26 3be1bf6
FORMAT: Introduce sub-function to clean-up _merge_two_compartments
