### 2024-04-26 06ace9a
FIX: Ensure that compartments have at least 2 connections after flow thresholding is applied
