### 2024-04-29 2ef5a7e
NEW: Consider flowrate when identifying compartments to merge
