### 2024-04-29 39b5a74
FIX: Change merge_illformed_compartments to use same criteria and selection as _merge_two_compartments
