### 2024-04-30 1df9004
FORMAT: Cleanup type hints
