### 2024-05-01 d88e220
NEW: Closed system example (OpenCMP)
