### 2024-05-02 ff85c03
NEW: Add element centroid to CMesh
