+++ openccm/compartment_models/pfr.py
@@ -132,36 +132,24 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
     # 1.4 Reorder connections and merge their locations as needed.
     connection_locations: Dict[int, List[Tuple[float, List[int]]]] = dict()
     for id_compartment, connections_distances_i in connection_distances.items():
-        compartment_connections = connection_pairing[id_compartment]
         # Check that the compartment connections are correct
+        compartment_connections = connection_pairing[id_compartment]
         _connections = np.array(list(compartment_connections.keys()))
         assert np.any(_connections > 0)
         assert np.any(_connections < 0)
 
-        # Re-organize inlet/outlet info in a way that is useful for merging
-        # Using this format so that the list can be sorted by distance
-        connections: List[Tuple[float, int]] = [(connections_distances_i[id], id) for id in connections_distances_i]
-
-        # Sort the list by distance (0th element) in preparation for merging entries that are too close
-        connections.sort(key=lambda element: element[0])
-
         # 1.4a If needed, add the artificial flows to the compartment
         if need_extra_flows:
             # Put in to be the second inlet into the compartment
-            i_first_inlet = next(i for i, (_, id_connection) in enumerate(connections) if id_connection > 0)
-            connections.insert(i_first_inlet+1, (connections[i_first_inlet][0], compartment_to_extra_connection[id_compartment]))
+            i_first_inlet = next(i for i, (_, id_connection) in enumerate(connections_distances_i) if id_connection > 0)
+            connections_distances_i.insert(i_first_inlet+1, (connections_distances_i[i_first_inlet][0], compartment_to_extra_connection[id_compartment]))
 
         # 1.4b Reorder connections
-        connections = _fix_connection_ordering(connections, _volumetric_flows, id_compartment, atol_opt)
-
+        connections_distances_i = _fix_connection_ordering(connections_distances_i, _volumetric_flows, id_compartment, atol_opt)
         # 1.4c Re-order domain inlets/outlets
-        connections = _fix_domain_boundary_connection_ordering(connections, compartment_connections, mesh.grouped_bcs, id_compartment)
-
+        connections_distances_i = _fix_domain_boundary_connection_ordering(connections_distances_i, compartment_connections, mesh.grouped_bcs, id_compartment)
         # 1.4d Merge connection locations
-        connections_merged = _merge_connections(connections, _volumetric_flows, dist_threshold, compartment_connections, atol_opt)
-
-        # Save the information about the connection locations
-        connection_locations[id_compartment] = connections_merged
+        connection_locations[id_compartment] = _merge_connections(connections_distances_i, _volumetric_flows, dist_threshold, compartment_connections, atol_opt)
 
     ####################################################################################################################
     # 2. Split compartment into PFRs
@@ -191,8 +179,7 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
     # The mapping must be made here but is only applies in the loop since the positive and negative signs
     # on the connection ID are still needed in order to split them into inlets/outlets.
     id_new = 0
-    ids_old = list(_volumetric_flows.keys())
-    ids_old.sort()
+    ids_old = sorted(_volumetric_flows.keys())
     map: Dict[int, int] = dict()
 
     for id_old in ids_old:
@@ -200,9 +187,8 @@ def create_pfr_network(compartments:        Dict[int, Set[int]],
         _volumetric_flows[id_new] = _volumetric_flows.pop(id_old)
         id_new += 1
 
-    connections: Dict[int, Tuple[Dict[int, int], Dict[int, int]]] = dict()
-
     # Re-numbing inlet and outlet information and adding it to the final dictionary
+    connections: Dict[int, Tuple[Dict[int, int], Dict[int, int]]] = dict()
[…]
