### 2024-05-08 8480626
NEW: Mapping each element to distance along PFR (or CSTR).
