### 2024-05-09 78c1dd7
NEW: Support OpenFOAM visualization output for PFR-based CMs
