+++ openccm/postprocessing/__init__.py
@@ -26,7 +26,8 @@ import numpy as np
 
 from ..config_functions import ConfigParser
 from ..mesh import CMesh
-from .vtu_output import cstrs_to_vtu_and_save_opencmp, pfrs_to_vtu_and_save_opencmp, cstrs_to_vtu_and_save_openfoam, \
+from .vtu_output import cstrs_to_vtu_and_save_opencmp, pfrs_to_vtu_and_save_opencmp, \
+                        cstrs_to_vtu_and_save_openfoam, pfrs_to_vtu_and_save_openfoam, \
                         create_element_label_gfu, create_compartment_label_gfu, label_compartments_openfoam
 from .analysis import network_to_rtd, plot_results, visualize_model_network
 
@@ -70,6 +71,6 @@ def convert_to_vtu_and_save(OpenCMP:        bool,
             cstrs_to_vtu_and_save_opencmp(system_results, compartments, config_parser, OpenCMP_mesh)
     else:
         if model == 'pfr':
-            raise NotImplementedError("PFR visualization for OpenFOAM input not yet implemented")
+            pfrs_to_vtu_and_save_openfoam(system_results, model_network, compartments, config_parser, cmesh)
         else:
             cstrs_to_vtu_and_save_openfoam(system_results, compartments, config_parser, cmesh)
