+++ openccm/postprocessing/__init__.py
@@ -26,10 +26,11 @@ import numpy as np
 
 from ..config_functions import ConfigParser
 from ..mesh import CMesh
+from .analysis import network_to_rtd, plot_results, visualize_model_network
 from .vtu_output import cstrs_to_vtu_and_save_opencmp, pfrs_to_vtu_and_save_opencmp, \
                         cstrs_to_vtu_and_save_openfoam, pfrs_to_vtu_and_save_openfoam, \
-                        create_element_label_gfu, create_compartment_label_gfu, label_compartments_openfoam
-from .analysis import network_to_rtd, plot_results, visualize_model_network
+                        create_element_label_gfu, create_compartment_label_gfu, \
+                        label_compartments_openfoam, label_elements_openfoam, label_models_and_dof_openfoam
 
 
 def convert_to_vtu_and_save(OpenCMP:        bool,
