### 2024-05-10 87577c4
NEW: Handle OpenFOAM visulzation for both PFR- and CSTR-based CMs with the same function
