+++ openccm/postprocessing/vtu_output.py
@@ -264,37 +264,23 @@ def label_compartments_openfoam(output_file_name: str, compartments: Dict[int, S
                 output_file.write(line)
 
 
-def pfrs_to_vtu_and_save_openfoam(
-        system_results:     Tuple[
-                            np.ndarray,
-                            np.ndarray,
-                            Dict[int, List[Tuple[int, int]]],
-                            Dict[int, List[Tuple[int, int]]]
-                        ],
-        pfr_network:    Tuple[
-                            Dict[int, Tuple[Dict[int, int], Dict[int, int]]],
-                            np.ndarray,
-                            np.ndarray,
-                            Dict[int, List[int]],
-                            List[List[Tuple[float, int]]]
-                        ],
-        compartments:   Dict[int, Set[int]],
-        config_parser:  ConfigParser,
-        cmesh:          CMesh
+def export_to_vtu_openfoam(
+        concentrations_all_time:    np.ndarray,
+        ts:                         np.ndarray,
+        model_to_element_map:       List[List[Tuple[float, int]]],
+        config_parser:              ConfigParser,
+        cmesh:                      CMesh
 ) -> None:
     print("Exporting simulation visualization")
 
     output_folder_path  = config_parser.get_item(['SETUP',              'output_folder_path'],  str)
     vtu_folder_path     = config_parser.get_item(['POST-PROCESSING',    'vtu_dir'],             str)
-    points_per_pfr      = config_parser.get_item(['SIMULATION',         'points_per_pfr'],      int)
+    points_per_model    = config_parser.get_item(['SIMULATION',         'points_per_pfr'],      int)
     species_names       = config_parser.get_list(['SIMULATION',         'specie_names'],        str)
 
     num_elements = len(cmesh.element_sizes)
 
-    concentrations_all_time, ts, _, _ = system_results
-    pfr_to_element_map = pfr_network[-1]
-
-    dof_to_element_map = create_dof_to_element_map(pfr_to_element_map, points_per_pfr)
+    dof_to_element_map = create_dof_to_element_map(model_to_element_map, points_per_model)
 
     assert len(dof_to_element_map) == concentrations_all_time.shape[1]
 
@@ -362,106 +348,6 @@ def write_buffer_to_file_openfoam(buffer: np.ndarray, output_file_path: str, t: 
             "\n"
             "// ************************************************************************* //"
         )
-
-
-def cstrs_to_vtu_and_save_openfoam(
-        system_results: Tuple[
-                            np.ndarray,
-                            np.ndarray,
-                            Dict[int, List[Tuple[int, int]]],
-                            Dict[int, List[Tuple[int, int]]]
-                        ],
-        compartments:   Dict[int, Set[int]],
-        config_parser:  ConfigParser,
-        cmesh:          CMesh) -> None:
-    """
-    Takes a time series from running a simulation on the compartment network and outputs it to native OpenFOAM format.
-    Each mesh cell is labelled based on which compartment it was a part of.
-    Cells which were not part of any compartment are labelled with -1.
-
-    Parameters
-    ----------
-    * system_results:   The simulation results to visualize (concentrations, times, inlets mapping, outlets mapping).
-    * compartments:     The compartments
-    * config_parser:    The OpenCCM ConfigParser.
-    * cmesh:            The CMesh from which the model and network were calculated.
-    """
-    print("Exporting simulation visualization")
-
-    output_folder_path          = config_parser.get_item(['SETUP',              'output_folder_path'],          str)
-    vtu_folder_path             = config_parser.get_item(['POST-PROCESSING',    'vtu_dir'],                     str)
-    species_names               = config_parser.get_list(['SIMULATION',         'specie_names'],                str)
-
-    element_to_compartment: Dict[int, int] = {}
-    for compartment_id, elements in compartments.items():
-        for element_id in elements:
[…]
