+++ openccm/postprocessing/__init__.py
@@ -28,7 +28,7 @@ from ..config_functions import ConfigParser
 from ..mesh import CMesh
 from .analysis import network_to_rtd, plot_results, visualize_model_network
 from .vtu_output import cstrs_to_vtu_and_save_opencmp, pfrs_to_vtu_and_save_opencmp, \
-                        cstrs_to_vtu_and_save_openfoam, pfrs_to_vtu_and_save_openfoam, \
+                        export_to_vtu_openfoam, \
                         create_element_label_gfu, create_compartment_label_gfu, \
                         label_compartments_openfoam, label_elements_openfoam, label_models_and_dof_openfoam
 
@@ -71,7 +71,4 @@ def convert_to_vtu_and_save(OpenCMP:        bool,
         else:
             cstrs_to_vtu_and_save_opencmp(system_results, compartments, config_parser, OpenCMP_mesh)
     else:
-        if model == 'pfr':
-            pfrs_to_vtu_and_save_openfoam(system_results, model_network, compartments, config_parser, cmesh)
-        else:
-            cstrs_to_vtu_and_save_openfoam(system_results, compartments, config_parser, cmesh)
+        export_to_vtu_openfoam(system_results[0], system_results[1], model_network[-1], config_parser, cmesh)
