### 2024-05-10 355643a
NEW: Simplify label_elements_openfoam
