### 2024-05-13 7b33013
NEW: Check BC and variable names against reserved keywords.
