### 2024-05-13 bfbd431
FORMAT: Change `solve_system` to take CMesh as an argument instead of GroupedBCs
