### 2024-05-14 d0b6357
NEW: Spatially varying initial conditions
