### 2024-05-23 2d986c4
NEW: Save CSTR- and PFR-based compartment networks as differently named files (they are created using different requirements).
