### 2024-05-29 df231ad
FORMAT: Change intra-compartment flowrate variable name to be more descriptive
