### 2024-05-30 3b9011a
FIX: Indexing error in `needs_merging`
