### 2024-05-31 #6576 @kyleniemeyer
Hi @philipcardiff, could you edit this? It involves modeling reacting flows, and integrates with OpenFOAM.
