### 2024-06-04 eff9e2a
FIX: Added missing else-statement
