### 2024-06-25 #6576 @alexrd
Hello,

Thanks for the invitation, but I am unable to review at this time.

Best,
Alex
---
Alex Dickson
Associate Professor
Department of Biochemistry and Molecular Biology
Department of Computational Mathematics, Science and Engineering
Michigan State University
603 Wilson Rd, Office 310B
East Lansing, MI  48824
***@***.*** ***@***.***>
http://www.egr.msu.edu/~alexrd/ <http://www.egr.msu.edu/~alexrd/>

> On Jun 25, 2024, at 10:46 AM, Philip Cardiff ***@***.***> wrote:
>
>
> Hello @alexrd <https://urldefense.com/v3/__https://github.com/alexrd__;!!HXCxUKc!0LWkjogfiVcnwQMICUmujty5Y1Pz671xi6ebxufJ75Q6lQwvwL7rdmCmXFlUP1mqRoDtPheu8ww4kdlXuuOoSsM$>, Would you be interested in reviewing the submission “OpenCCM: An Open-Source Continuous Compartmental Modelling Package”, which we received at JOSS? The link for the submission is #6576 <https://urldefense.com/v3/__https://github.com/openjournals/joss-reviews/issues/6576__;!!HXCxUKc!0LWkjogfiVcnwQMICUmujty5Y1Pz671xi6ebxufJ75Q6lQwvwL7rdmCmXFlUP1mqRoDtPheu8ww4kdlXXXRWr2U$>, and the repository is at https://github.com/uw-comphys/openccm/tree/main <https://urldefense.com/v3/__https://github.com/uw-comphys/openccm/tree/main__;!!HXCxUKc!0LWkjogfiVcnwQMICUmujty5Y1Pz671xi6ebxufJ75Q6lQwvwL7rdmCmXFlUP1mqRoDtPheu8ww4kdlXoz1JsWY$>.
>
> If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/review_criteria.html__;!!HXCxUKc!0LWkjogfiVcnwQMICUmujty5Y1Pz671xi6ebxufJ75Q6lQwvwL7rdmCmXFlUP1mqRoDtPheu8ww4kdlXukonXwQ$>) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/review_checklist.html__;!!HXCxUKc!0LWkjogfiVcnwQMICUmujty5Y1Pz671xi6ebxufJ75Q6lQwvwL7rdmCmXFlUP1mqRoDtPheu8ww4kdlXlMhACiU$>) to get an idea of what is expected.
>
