### 2024-06-26 #6576 @philipcardiff
Hello @luohancfd, Would you be interested in reviewing the submission “OpenCCM: An Open-Source Continuous Compartmental Modelling Package”, which we received at JOSS? The link for the submission is https://github.com/openjournals/joss-reviews/issues/6576, and the repository is at https://github.com/uw-comphys/openccm/tree/main.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.
