### 2024-06-26 ca6caa7
FIX: Correctly checking that each connection is used properly.
