### 2024-06-26 52d03a4
NEW: Check that each connection is used at least once.
