### 2024-06-27 151471f
NEW: Check for disconnected subgraphs
