### 2024-07-04 #6576 @speth
Hi @philipcardiff. Sure, I'd be happy to.
