### 2024-07-05 #6576 @philipcardiff
> Hi @philipcardiff. Sure, I'd be happy to.

Thanks, @speth.
