### 2024-07-15 uw-comphys/openccm#37 @luohancfd
JOSS review
Hi authors! I have a few comments regarding to the [review](https://github.com/openjournals/joss-reviews/issues/6963).

- [x] The package `NetworkX` should be added as a required package
- [x] Can you provide a case similar to the extra material at https://www.sciencedirect.com/science/article/pii/S0098135424000681?via%3Dihub that calculates RTD? In other words, the script that reproduces Figure 4 in the JOSS paper. I'm especially interested in seeing a case based on OpenFOAM flow results.
- [x] Minor edit: the code needs Administrator privilege to run on Windows due to `os.symlink`. It might be good to throw an error if the requirement is not met at the beginning.
