### 2024-07-15 uw-comphys/openccm#37 @Alex-Vasile
**NetworkX**
The package [is currently listed](https://github.com/uw-comphys/openccm/blob/73f4f0a11d75006df7b59d8bd9f786284753a48c/setup.cfg#L49) as an optional requirement as it isn't needed for the main functionality of the package. It's only needed for certain optional post processing functionality, specifically if `network_diagram` is set to `True`, which is the current default.
- [ ] Do you still get a package not found error if `network_diagram` is set to `False` in the config file? (It goes under `POST-PROCESSING`). If so, this is an oversight on my part as the imports were supposed to be wrapped so that they would only be called when that flag was used.
- [ ] Would changing the provided examples, either by overriding `network_diagram` in their config file or by changing the default value, such that they do not use the NetworkX code path be an acceptable solution?

**RTD**
I will look into incorporating that calculation script into the OpenCMP example provided, as that example and the simulation from the paper are derived from the same source.

**Symlink**
I will take a look and test alternatives for symlinking. I'm thinking automatically falling back to a plain file copy, and displaying a warning, would be preferable to throwing the error.

This would let the user know that a faster alternative is available, but does not force them to run with admins privileges, which may be an impossible expectation in certain corporate engineering settings. And it would generally be less risky for the user to not have to run it with admin privileges.
