### 2024-08-09 f14774a
FIX: Disable networkX usage by default since it's an optional dependency
