### 2024-08-14 8555826
FIX: Remove need for admin privileges for Windows
