### 2024-08-14 b35072c
NEW: Update OpenCMP example to recreate figure 4
