+++ examples/OpenCMP/pipe_with_recirc_2d/analysis.py
@@ -1,0 +1,221 @@
+from shutil import rmtree
+from typing import List, Tuple, Dict, Optional
+from pathlib import Path
+
+import pickle
+
+import numpy as np
+import scipy as sp
+import matplotlib.pyplot as plt
+
+
+import ngsolve as ngs
+import pyngcore as ngcore
+from netgen.read_gmsh import ReadGmsh
+from ngsolve import GridFunction, Mesh, Integrate
+
+from openccm.config_functions import ConfigParser
+from openccm.io import fes_from_opencmp_config
+from openccm import run
+
+fig_width_pt = 390.0 * 1.7  # Get this from LaTeX using \showthe\columnwidth
+inches_per_pt = 1.0 / 72.27  # Convert pt to inch
+golden_mean = 2.0 / (np.sqrt(5) + 1.0)  # Aesthetic ratio
+fig_width = fig_width_pt * inches_per_pt  # width in inches
+fig_height = fig_width * golden_mean  # height in inches
+fig_size = [fig_width, fig_height]
+params = {'axes.labelsize': 18,
+          'font.size': 18,
+          'legend.fontsize': 18,
+          'xtick.labelsize': 18,
+          'ytick.labelsize': 18,
+          'figure.figsize': fig_size}
+plt.rcParams.update(params)
+
+linewidth = 4
+
+xlim = [0, 300]
+
+
+def sort_folders_by_name(folders: List[Path], all_int: bool = False) -> List[Path]:
+    """
+    Helper function to deal with sorting folders by name in numerical order when they are floats
+    """
+    # If running on a mac, need to make sure to delete and .DS_Store files
+    i = 0
+    while i < len(folders):
+        folder = folders[i]
+        if folder.name == '.DS_Store':
+            folders.pop(i)
+        else:
+            if all_int:
+                assert folder.name.isalnum()
+            i += 1
+    if all_int:
+        folders_info = [(int(folder.name), folder) for folder in folders]
+    else:
+        folders_info = [(float(folder.name.split("_")[1][:-4]), folder) for folder in folders]
+    folders_info.sort(key=lambda tup: tup[0])
+
+    return [folder_info[1] for folder_info in folders_info]
+
+
+def rtd_cfd() -> np.ndarray:
+    """
+    Function to calculate E(t) and F(t) for the CFD tracer experiments without converting to .vtu.
+    """
+    # Get all the results folders
+    folder = Path(".")
+
+    # Set parameters for ngsolve
+    ngcore.SetNumThreads(4)
+
+    # Run the model.
+    with (ngcore.TaskManager()):
+        # Get data info from file
+        msh_files = [file for file in folder.glob('*.msh')]
+        assert len(msh_files) == 1
+        config_files = [file for file in folder.glob('config_MCINS')]
+        assert len(config_files) == 1
+        config_file_path = str(config_files[0])
+
+        # Load mesh and create grid function
+        data_path               = Path(str(folder.absolute()) + "/cfd_rtd.npy")
+        if data_path.exists():
+            rtd_cfd = np.load(str(data_path.absolute()))
+        else:
+            mesh = Mesh(ReadGmsh(str(msh_files[0])))
+            fes = fes_from_opencmp_config(config_file_path, mesh)
+            gfu = GridFunction(fes)
+            u: GridFunction = gfu.components[0]  # Hardcoded
+            tracer: GridFunction = gfu.components[2]  # Hardcoded
+
+            # Get list of .sol files
[…]
