### 2024-08-15 uw-comphys/openccm#37 @Alex-Vasile
@luohancfd Hi there, let me know if the changes in #38 properly address your above comments.
