### 2024-08-17 uw-comphys/openccm#40 @speth
Missing installation instructions
I could not find any installation instructions. While I found that running the common `pip install -e .` from a Git checkout worked, even this installation route should be documented.

- There is nothing mentioned in `README.md`
- There is no separate `INSTALL.md`
- There is no information available via the linked documentation website (https://uw-comphys.github.io/openccm/)
- There is no "release" version marked on GitHub, and the only tagged version is v0.9. Meanwhile, the version available from PyPI is v0.1.
- The method for installing with the extra dependencies for the examples should be documented. I had to read `setup.cfg` and search Stack Overflow to learn that the correct command is `pip install -e ".[all]"` .

This is in the context of the review for JOSS (https://github.com/openjournals/joss-reviews/issues/6963).
