### 2024-08-17 uw-comphys/openccm#41 @speth
Test suite fails
Most of the tests fail when I run them (`main` branch, commit 8ba12ae).

First, I note that the correct way to run the tests does not appear to be documented -- it requires running `pytest` from the `pytests` subdirectory, and does not work if you run `pytest pytests/` from the root directory of the source tree.

The short summary of failures is as follows:

```
=============================== short test summary info ================================
FAILED test_run_examples.py::test_opencmp_cstr_reversible - NameError: name 'np' is not defined
FAILED test_run_examples.py::test_opencmp_cstr_irreversible - NameError: name 'np' is not defined
FAILED test_run_examples.py::test_opencmp_pfr - NameError: name 'np' is not defined
FAILED test_run_examples.py::test_opencmp_recirc - NameError: name 'np' is not defined
FAILED test_run_examples.py::test_openfoam_2d_pipe - ValueError: Specified folder (../examples/OpenFOAM/pipe_with_recirc/./) does not co...
======================== 5 failed, 1 passed, 1 warning in 3.13s ========================
```

The full output of `pytest -v .` is here:[testlog.txt](https://github.com/user-attachments/files/16645716/testlog.txt)

I am not sure what the requirements for running the test suite are. Does it require that both OpenCMP and OpenFOAM are also installed? It might be better to include some small example input data that can be used for testing, rather than requiring a rather large dependency to be installed.

This is in the context of the review for JOSS (https://github.com/openjournals/joss-reviews/issues/6963).
