### 2024-08-17 uw-comphys/openccm#42 @speth
Documentation needs more structure and details
The API documentation at [uw-comphys.github.io/openccm/](https://uw-comphys.github.io/openccm/) could use some improvements to make relevant capabilities more obvious to the user.

Currently, the landing page and most of the pages directly accessible from there are either blank (https://uw-comphys.github.io/openccm/openccm/config_functions.html) or just provide function signatures for undocumented functions (https://uw-comphys.github.io/openccm/openccm/compartment_models.html).

It is completely unclear what the main user "entry points" for using the software are, and there are no links to any examples or tutorials.

Some attention should be given to how the docstrings written in the code are actually rendered in HTML. For example, there is a lot of structure given in the source for `create_pfr_network`, but in the HTML docs (https://uw-comphys.github.io/openccm/openccm/compartment_models/pfr.html#create_pfr_network) most of this is lost in long, mushed together paragraphs. The same is true for a number of other functions.

There are also no links back to the GitHub repository or other resources for users who want to contribute, report issues, or need help installing / using the software.

This is in the context of the review for JOSS (https://github.com/openjournals/joss-reviews/issues/6963).
