### 2024-08-17 uw-comphys/openccm#43 @speth
JOSS Review - comments on paper
I appreciate the opportunity to review this tool for JOSS (https://github.com/openjournals/joss-reviews/issues/6963). OpenCCM seems like it could be a useful tool in a number of different areas. Besides the other issues I opened regarding the software, I wanted to offer some comments / suggestions for the paper itself:

- Summary, L20: OpenFOAM reference is undated, but the linked doc does have a date that could be used
- The in-text reference to AMBER is awkward, appearing as "(Compartment Modelling Software, n.d.)" -- my suggestion would be to list the name of the company as the "Author".
- In the "Features" table:
  - It is a bit hard to tell what "Description" lines go with what "Feature", especially the line "Previous SotA CSTR-based models". I'm not sure how easy this is to improve in a Markdown table, though.
  - The acronyms PFR, CSTR here are not expanded even though this is the first time they are used
  - "state of the art" should just be written out.
- In the description of the user interface, the input file is described as "text based," but is there any more structure to it than that? An example would aid in understanding.
- L78: "specie" should be "species"
- L82/L117: The chemical equation is poorly formatted -- elements should not be italicized, and the subscripts should be subscripts.
- L118: What are the units of the rate constants specified here. And later, the units of concentration?
