### 2024-08-19 66f10a9
FIX: Round dot product to address floating-point rounding issues for arccos
