### 2024-08-19 ba857e8
FIX: Change BC generation code to work with numpy 2.0
