### 2024-08-20 uw-comphys/openccm#39 @Alex-Vasile
Hi @speth, thanks for the thorough message. I will have a PR to address these issues up shortly, in the meantime I shall comment on them here:

- [x] Syntax warning is an unescaped "\". That's been fixed.
- [x] The invalid value warning in the arccos was caused by numerical imprecission; sometimes it would get values of 1.000000000002 which should really be 1.0. This has been addressed by rounding.
- [x] The physical groups warning is done in NGSolve (OpenCMP's finite element library), it's not a sign that something is wrong or has been missed, it's letting the user know that by using Physical groups they have to be careful since the code itself is unable validate the choices made, or more specifically any missing values. There is nothing wrong with this warning.
- [x] The fatal numpy error is produced by our code gen. When numpy when from 1.2 to 2.0 it changed how it represented floats when turned into a string; in numpy 1.2 the np.float64(...) was not there and it was only the number. This has been fixed by adding an import statement to the code gen.
- [x] The subsequent run issues. Yes, this was poorly coded on my part and was not resilient to the `run` call failiing. I have improved that and this has been fixed.
- [ ] The two leftover warnings ("a large number ..." and "merged 98.4855% ...") are similar to the physical group warning above; something is not necessarily wrong, but the software cannot discern that and it instead flags it for the user. I can tweak these settings to get rid of these warnings.
