### 2024-08-20 a18be55
FIX: re-include steady-state velocity profile for OpenFOAM example
