### 2024-08-20 71fbc36
FIX: Off-by-one error in checks
