### 2024-08-20 88224ec
FIX: Increased tolerance for warning the user about merging compartments.
