### 2024-08-20 uw-comphys/openccm#39 @Alex-Vasile
Hi @speth, please let me know if #44 addresses your issues.
