### 2024-08-20 af57657
FIX: Add run instructions
