### 2024-08-20 98bd9e6
FIX: Auto generate OpenCMP data for pipe_with_recirc example for tests.
