### 2024-08-20 0caa40c
FIX: Skip tests that require OpenCMP if it's not installed
