### 2024-08-20 uw-comphys/openccm#41 @Alex-Vasile
Hi @speth.

I'll be putting up a PR for this soon, see my comments below.

- [x] I have included instructions specifying where to run the tests from.
- [x] Most of the failures are due to the a numpy change when going from 1.2 to 2.0. #39 addresses it.
- [x] The OpenFOAM failure was caused by unavailable velocity profiles. I have included one and modified the run instructions for it accordingly.
- [x] I have also included code for auto-running OpenCMP for generating the data needed for the tests when they are run.
- [x] Re: Requirements. For the currently available OpenFOAM tests I have included the velocity data required in native OpenFOAM format, but do not want to generally do this as the file sizes will quickly get out of hand. In order to run all of the tests OpenCMP is required to be installed. I have changed the OpenCMP tests to be skipped if OpenCMP is not installed if the end user only cares about OpenFOAM support. The need for OpenCMP installed for it's test are:
1. OpenCMP functionality is leveraged for file I/O.
2. We have considered including OpenCMP in our intermediary format, but that would require re-processing whenever the intermediary format or the conversion process is changed. Further,  the tests would no longer be hitting the conversion steps. Finally it would increase the size of the repository and of the final installed package.
