+++ publications/2403_JOSS/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 `OpenCCM` is a compartmental modelling [@Jourdan2019] software package which is based on recently-developed fully automated flow alignment compartmentalization methods [@Vasile2024]. It is primarily intended for large-scale flow-based processes where there is weak coupling between composition changes, e.g. through (bio)chemical reactions, and convective mass transport in the system. Compartmental modelling is an important approach used to developed reduced-order models [@Chinesta2017] [@Benner2020] using a priori knowledge of process hydrodynamics [@Jourdan2019]. Compartmental modelling methods, such as those implemented in `OpenCCM`, enable simulations of these processes to be performed with far less computational complexity while still capturing the key aspects of their performance.
 
-`OpenCCM` integrates with two multiphysics simulation software packages, `OpenCMP` [@Monte2022] and `OpenFOAM` [@OpenFOAM], allowing for ease of transferring simulation data for compartmentalization. Additionally, it provides users with built-in functionality for calculating residence times, exporting to transfer data to simulation software, and for exporting results for visualization using `ParaView` [@Paraview]. Post-processing methods are included for mapping simulation results from compartment domains to the original simulation domain, useful for both visualization purposes and for further simulations in other software (e.g. multi-scale modelling).
+`OpenCCM` integrates with two multiphysics simulation software packages, `OpenCMP` [@Monte2022] and `OpenFOAM` [@greenshields2024], allowing for ease of transferring simulation data for compartmentalization. Additionally, it provides users with built-in functionality for calculating residence times, exporting to transfer data to simulation software, and for exporting results for visualization using `ParaView` [@Paraview]. Post-processing methods are included for mapping simulation results from compartment domains to the original simulation domain, useful for both visualization purposes and for further simulations in other software (e.g. multi-scale modelling).
 
 
 # Statement of Need
@@ -50,22 +50,27 @@ The overall aim of `OpenCCM` is to fill this need for an open-source compartment
 
 # Features
 
-| Feature                 | Description                                                                         |
+| **FEATURE**             | **DESCRIPTION**                                                                     |
 |-------------------------|-------------------------------------------------------------------------------------|
-| Model support           | Accepts `OpenCMP` [@Monte2022] and `OpenFOAM` [@OpenFOAM] results                   |
+| Model support           | Accepts `OpenCMP` [@Monte2022] and `OpenFOAM` [@greenshields2024] results           |
+|-------------------------| ------------------------------------------------------------------------------------- |
 | Compartmentalization    | Single-phase flow-based compartment identification                                  |
-| Compartmental Modelling | PFR-in-series-based model                                                           |
-|                         | Previous SotA CSTR-based models                                                     |
+|-------------------------| ------------------------------------------------------------------------------------- |
+| Compartmental Modelling | Plug Flow Reactors (PFRs)-in-series-based model                                     |
+|                         | Previous state-of-the-art Continous Stirred-tank reactor (CSTR)-based models        |
+|-------------------------| ------------------------------------------------------------------------------------- |
 | CM Simulations          | Linear, non-linear, and reversible arbitrary reactions.                             |
 |                         | 1st Order upwinding finite-difference-based                                         |
 |                         | Adaptive time-stepping                                                              |
+|-------------------------| ------------------------------------------------------------------------------------- |
 | Post-Processing         | Residence time distribution                                                         |
[…]
