+++ publications/2403_JOSS/paper.bib
@@ -11,18 +11,21 @@
 
 @misc{Amber,
   title = {Compartment Modelling Software},
+  author = {Quintessa},
   howpublished = {Online},
   organization = {Quintessa},
   url = {https://www.quintessa.org/software/AMBER},
   urldate = {2024-01-03}
 }
 
-@misc{OpenFOAM,
-  title = {OpenFOAM v10 User Guide},
-  howpublished = {Online},
-  organization = {Open{FOAM} Foundation},
-  url = {https://doc.cfd.direct/openfoam/user-guide-v10/index},
-  urldate = {2024-01-03}
+@book{
+  greenshields2024,
+  title     = "OpenFOAM v12 User Guide",
+  author    = "Greenshields, Christopher",
+  year      = 2024,
+  url       = "https://doc.cfd.direct/openfoam/user-guide-v12",
+  publisher = "The OpenFOAM Foundation",
+  address   = "London, UK"
 }
 
 @article{Monte2022,
