### 2024-08-21 5e090fc
FIX: Update reaction figures
