### 2024-08-21 7e709bb
FIX: Update figure based on new config values
