### 2024-08-21 09cefe5
FORMAT: Units, better formatting of equations, and specifying correct limitations for reaction parser
