+++ publications/2403_JOSS/paper.md
@@ -93,11 +93,11 @@ The chemical reaction parser in `OpenCCM` reads and parses the reactions configu
 
 `aA + bB + [...] -> cC + dD + [...]`
 
-with associated numeric rate constants. It intentionally does not support the standard `<->` symbol for reversible chemical reactions, so that each independent reaction has an explicit rate constant clearly defined in the same file. Therefore, a reversible reaction must be written as two independent forward reactions (with separate rate constants). Each species *label* must solely contain letter characters, e.g. `O` instead of `O2` for oxygen. Kinetic rate constants must be expressed as positive real numbers in standard or scientific notation. Additionally, each reaction/rate pair must also have a unique *identifier* (i.e. R1, R2). For example, take the reversible reaction,
+with associated numeric rate constants. It intentionally does not support the standard `<->` symbol for reversible chemical reactions, so that each independent reaction has its rate constant explicit defined. Therefore, a reversible reaction must be written as two independent forward reactions, each with its own rate constants. Each species *label* can contain letter and numbers, but cannot contain brackets "(" or ")" or special characters, e.g. "+", "-", "^", etc. Kinetic rate constants must be expressed as positive real numbers in standard or scientific notation. Additionally, each reaction/rate pair must have a unique *identifier* (i.e. R1, R2). For example, take the reversible reaction,
 
 $$2\textrm{NaCl} + \textrm{CaCO}_3 \Leftrightarrow \textrm{Na}_2\textrm{CO}_3 + \textrm{CaCl}_2$$
 
-with `k_f = 5e-2` and `k_r = 2` the species must first be redefined in simple terms in agreement with the reactions parser, i.e. a = NaCl, b = CaCO3, c = Na2CO3, and d = CaCl2. A configuration file for this reversible reaction may then be:
+with $k_f = 5e-2$ and $k_r = 2$, dimensionless numbers picked for demonstration purposes. A configuration file for this reversible reaction may then be:
 
     [REACTIONS]
     R1: 2NaCl   + CaCO3 ->  Na2CO3 + CaCl2
@@ -107,11 +107,11 @@ with `k_f = 5e-2` and `k_r = 2` the species must first be redefined in simple te
     R1: 5e-2
     R2: 2
 
-where **R1** and **R2** are the reaction *identifiers* for the forward and reverse reactions respectively.
+where **R1** and **R2** are the reaction *identifiers* for the forward and reverse reactions, respectively.
 
 # Examples of Usage
 
-Several examples are provided in the `OpenCCM` documentation which demonstrate the usage of both `OpenCMP` and `OpenFOAM` simulation flow information for compartmentalization. One example uses the geometry from [@Vasile2024] and shows how to execute the needed CFD simulation for flow information (both using `OpenCMP` and `OpenFOAM`), create/visualize the compartmental model results, and compare the predicted RTD  to the reference result directly from CFD simulation.
+Several examples are provided in the `OpenCCM` documentation which demonstrate the usage of both `OpenCMP` and `OpenFOAM` simulation flow information for compartmentalization. One example, inside `examples/OpenCMP/pipe_with_recird_2d`, uses the geometry from [@Vasile2024] and shows how to execute the needed CFD simulation for flow information (both using `OpenCMP` and `OpenFOAM`), create/visualize the compartmental model results, and compare the predicted RTD  to the reference result directly from CFD simulation.
[…]
