### 2024-08-22 56dd815
FORMAT: Document the main __init__.py
