### 2024-08-22 uw-comphys/openccm#41 @Alex-Vasile
Hi @speth , let me know if the changes from #45 address your comments.
