### 2024-08-22 uw-comphys/openccm#43 @Alex-Vasile
> In the description of the user interface, the input file is described as "text based," but is there any more structure to it than that? An example would aid in understanding.

The main directory contains two example config files, `CONFIG` and `CONFIG_REACTIONS`. Would indicating their existence and pointing the reader to it address this comment?
