+++ openccm/mesh/cmesh.py
@@ -15,6 +15,11 @@
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
 
+r"""
+The intermediary mesh format, CMesh, used by all internal OpenCCM calculations so that the compartmentalization and simulation
+algorithms provided in this package are CFD-package agnostic.
+"""
+
 from typing import Tuple, Dict, List
 
 import numpy as np
@@ -24,50 +29,73 @@ from openccm.config_functions import ConfigParser
 
 class GroupedBCs:
     """
-    Helper class which handles the internal numbering of the BCs for the mesh.
+    Helper class which handles the internal numbering of the boundary conditions for the mesh.
     """
     def __init__(self, config_parser: ConfigParser):
-        no_flux_names       = config_parser.get_expression(['INPUT', 'no_flux_names'])
-        ignored_names       = config_parser.get_expression(['INPUT', 'ignored_names'])
+        """
+        Only initializer for  the class.
+
+        Parameters
+        ----------
+        * config_parser: The OpenCCM configuration parser which contains the
+        """
         domain_inlet_names  = config_parser.get_expression(['INPUT', 'domain_inlet_names'])
         domain_outlet_names = config_parser.get_expression(['INPUT', 'domain_outlet_names'])
+        ignored_names       = config_parser.get_expression(['INPUT', 'ignored_names'])
+        no_flux_names       = config_parser.get_expression(['INPUT', 'no_flux_names'])
 
-        assert (not no_flux_names)       or len(no_flux_names)       == len(set(no_flux_names))
-        assert (not ignored_names)       or len(ignored_names)       == len(set(ignored_names))
-        assert (not domain_inlet_names)  or len(domain_inlet_names)  == len(set(domain_inlet_names))
-        assert (not domain_outlet_names) or len(domain_outlet_names) == len(set(domain_outlet_names))
+        assert (not domain_inlet_names)  or (len(domain_inlet_names)  == len(set(domain_inlet_names)))
+        assert (not domain_outlet_names) or (len(domain_outlet_names) == len(set(domain_outlet_names)))
+        assert (not ignored_names)       or (len(ignored_names)       == len(set(ignored_names)))
+        assert (not no_flux_names)       or (len(no_flux_names)       == len(set(no_flux_names)))
 
-        self.no_flux_names:       Tuple[str, ...] = no_flux_names                      if no_flux_names       else tuple()
-        self.ignored_names:       Tuple[str, ...] = ignored_names                      if ignored_names       else tuple()
         self.domain_inlet_names:  Tuple[str, ...] = tuple(sorted(domain_inlet_names))  if domain_inlet_names  else tuple()
-        self.domain_outlet_names: Tuple[str, ...] = tuple(sorted(domain_outlet_names)) if domain_outlet_names else tuple()
-        self.domain_in_out_names: Tuple[str, ...] = tuple(inout for inout in self.domain_inlet_names + self.domain_outlet_names)
+        """The names of all boundaries which are inlets to the domain."""
+        self.domain_outlet_names: Tuple[str, ...] = tuple(
+            sorted(domain_outlet_names)) if domain_outlet_names else tuple()
+        """The names of all boundaries which are outlets to the domain."""
+        self.domain_in_out_names: Tuple[str, ...] = tuple(
+            inout for inout in self.domain_inlet_names + self.domain_outlet_names)
+        """The name of all boundaries which are inlets or outlets to the domain."""
+        self.ignored_names: Tuple[str, ...] = ignored_names if ignored_names else tuple()
+        """
+        The names of all boundaries which are ignored by the algorithm.
+        An example of such a boundary are the ignored boundaries used in OpenFOAM for creating a quasi-2D domain by creating a mesh
+        with a single layer of cells.
 
[…]
