+++ openccm/mesh/__init__.py
@@ -15,6 +15,13 @@
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
 
+r"""
+The mesh module contains the intermediate mesh representation, CMesh, and the functions required to convert OpenFOAM
+and NGSolve meshes to CMesh format.
+"""
+
+from typing import Optional
+
 import numpy as np
 
 from .cmesh import CMesh, GroupedBCs
@@ -23,29 +30,42 @@ from .convert_openfoam import convert_mesh_openfoam
 from ..config_functions import ConfigParser
 
 
-def convert_mesh(OpenCMP: bool, config_parser: ConfigParser, **kwargs) -> CMesh:
-    if OpenCMP:
-        # Done in order to keep NGSolve & OpenCMP dependency optional
-        from .convert_ngsolve import convert_mesh_ngsolve
-        return convert_mesh_ngsolve(config_parser, kwargs['mesh'])
-    else:
+def convert_mesh(config_parser: ConfigParser, ngsolve_mesh: Optional['ngsolve.Mesh']) -> CMesh:
+    """
+
+
+    Parameters
+    ----------
+    * config_parser:    The OpenCCM ConfigParser from which to get the required info for conversion.
+    * ngsolve_mesh:     The NGSolve mesh object to convert if using OpenCMP.
+
+    Returns
+    -------
+    * cmesh: Internal representation of the mesh inside ngsolve_mesh or pointed to by config_parser
+    """
+    if ngsolve_mesh is None:
         return convert_mesh_openfoam(config_parser)
+    else:
+        from .convert_ngsolve import convert_mesh_ngsolve  # Import here to NGSolve & OpenCMP dependencies optional
+        return convert_mesh_ngsolve(config_parser, ngsolve_mesh)
 
 
 def convert_velocities_to_flows(cmesh: CMesh, vel_vec: np.ndarray) -> np.ndarray:
     """
-    Convert cell-centered velocities to volumetric flowrate through facets using upwinding.
+    Convert cell-centered velocities to flowrate through facets using upwinding.
 
-    Args:
-        cmesh:      The CMesh to project on.
-        vel_vec:    Velocity vector indexed by element ID.
+    Parameters
+    ----------
+    * cmesh:      The CMesh to project on.
+    * vel_vec:    Velocity vector indexed by element ID.
 
-    Return:
-        flows_and_upwind: 2D object array indexed by facet ID.
-                            - 1st column is volumetric flowrate through facet.
-                            - 2nd column is a flag indicating which of a facet's elements are upwind of it.
-                                - 0, and 1 represent the index into mesh.facet_elements[facet]
-                                - -1 is used for boundary elements to represent
+    Returns
+    -------
+    * flows_and_upwind: 2D object array indexed by facet ID.
+                        - 1st column is volumetric flowrate through facet.
+                        - 2nd column is a flag indicating which of a facet's elements are upwind of it.
+                            - 0, and 1 represent the index into mesh.facet_elements[facet]
+                            - -1 is used for boundary elements to represent
     """
     flow = cmesh.facet_size.copy()
     upwind_element = np.zeros(len(flow), dtype=int)
