### 2024-08-23 ecd6e08
FORMAT: Documentation for compartment_models module
