+++ openccm/compartment_models/__init__.py
@@ -14,6 +14,11 @@
 # You should have received a copy of the GNU Lesser General Public License along with OpenCCM. If not, see             #
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
+
+r"""
+All functions related to take a network of compartments and convert them into a network of CSTRs or PFRs.
+"""
+
 from typing import Dict, Tuple, List, Set
 
 import numpy as np
@@ -37,6 +42,25 @@ def create_model_network(model:                 str,
             np.ndarray,
             Dict[int, List[int]]
         ]:
+    """
+    Helper function to clean up call sites.
+    Calls `openccm.compartment_models.pfr.create_pfr_network` or `openccm.compartment_models.cstr.create_cstr_network`
+    based on the value of `model`.
+
+    Parameters
+    ----------
+    * model:                The kind of model to use for building the network, 'pfr' or 'cstr'.
+    * compartments:         The element IDs that make up each compartment, indexed by compartment ID.
+    * compartment_network:  The compartment_network created from the compartments.
+    * mesh:                 The CMesh from which the compartments were built.
+    * dir_vec:              The direction vector in each mesh element, indexed by element ID.
+    * flows_and_upwind:     The flow through each ID and indicator of which way is upwind, indexed by facet ID.
+    * config_parser:        OpenCCM ConfigParser to use.
+
+    Returns
+    -------
+    * model_network: The created network object.
+    """
     if model == 'pfr':
         return create_pfr_network(compartments, compartment_network, mesh, flows_and_upwind, dir_vec, config_parser)
     else:
