+++ openccm/compartment_models/helpers.py
@@ -14,6 +14,11 @@
 # You should have received a copy of the GNU Lesser General Public License along with OpenCCM. If not, see             #
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
+
+r"""
+Helper functions common to both CSTR- and PFR-based networks.
+"""
+
 from typing import Dict, Set, Tuple
 
 import numpy as np
@@ -29,12 +34,13 @@ def check_network_for_disconnected_subgraphs(connection_pairings: Dict[int, Dict
 
     Throws a ValueError if there are disconnected subgraphs.
 
-    Args:
-        connection_pairings:    Mapping between compartment ID and it's neighbours.
-                                Each entry is another mapping between the connection ID and the neibhour ID.
-
-    Returns:
-          ~: None.
+    Parameters
+    ----------
+    * connection_pairings:  Mapping between compartment ID and its neighbours.
+                            Each entry is another mapping between the connection ID and the neighbour ID.
+    Returns
+    -------
+    * Nothing, will through an error if there are disconnected subgraphs.
     """
     uncolored_nodes = set(connection_pairings.keys())
     num_subgraphs = 0
@@ -66,62 +72,67 @@ def tweak_compartment_flows(
     - This method works balances the mass around each COMPARTMENT, that one balances the mass around each CSTR/PFR.
     - That method tries to minimize the error in the conservation of mass, accepting small deviations resulting in either
       a small net inflow or outflow.
-      This method accepts looser tolerances in order to ensure there is NEVER a net outflow.
+    - This method accepts looser tolerances in order to ensure there is NEVER a net outflow.
       This is requirement is needed by the Compartment -> PFR method, having even a small net outflow will result in the
       intra-compartment connections that get made to cause a reversal of flow.
 
     The first mass balance inequality written around each PFR is to ensure that there is a small netflow (epsilon)
     To ensure it's positive even after any small floating point precision issues.
-    Σ_inlets (Q + x) - Σ_outlets(Q + x) >= eps
-    (ΣQ_in - ΣQ_out) + (Σx_in - Σx_out) >= eps
-    (ΣQ_in - ΣQ_out)                    >= eps  - (Σx_in - Σx_out)
-    (-Σx_in + Σx_out)                   <= -eps + (ΣQ_in - ΣQ_out)
+
+        Σ_inlets (Q + x) - Σ_outlets(Q + x) >= eps
+        (ΣQ_in - ΣQ_out) + (Σx_in - Σx_out) >= eps
+        (ΣQ_in - ΣQ_out)                    >= eps  - (Σx_in - Σx_out)
+        (-Σx_in + Σx_out)                   <= -eps + (ΣQ_in - ΣQ_out)
 
     The second mass balance inequality is to put an upper bound on the net flow into the compartment
-    Σ_inlets (Q + x) - Σ_outlets(Q + x) <= atol_opt
-    (ΣQ_in - ΣQ_out) + (Σx_in - Σx_out) <= atol_opt
-    (Σx_in - Σx_out)                    <= atol_opt - (ΣQ_in - ΣQ_out)
+
+        Σ_inlets (Q + x) - Σ_outlets(Q + x) <= atol_opt
+        (ΣQ_in - ΣQ_out) + (Σx_in - Σx_out) <= atol_opt
+        (Σx_in - Σx_out)                    <= atol_opt - (ΣQ_in - ΣQ_out)
 
     Where:
-        - Q is the current flowrate through a connection
-        - x is the non-negative adjustment to a given connection
-        - eps is a small value used to avoid floating point issues when calculating the net flow.
-        - atol_opt is the absolute tolerance for the conservation of mass, the maximum allowable net inflow.
[…]

[…]
