### 2024-08-23 c5147a2
FORMAT: Documentation for compartmentalize module
