+++ openccm/compartmentalize/unidirectional.py
@@ -15,6 +15,10 @@
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
 
+r"""
+All functions related to identifying compartments as regions of unidirectional flow.
+"""
+
 from typing import Dict, List, Set, Tuple
 from collections import OrderedDict
 import numpy as np
@@ -43,22 +47,28 @@ def create_compartment_network(compartments:        Dict[int, Set[int]],
 
     Finally, the compartments are renumbered to the range [0, N) in a sequential order without gaps.
 
-    Args:
-        compartments:       A list of set. The ith set contains the indices of the elements belonging to the ith compartment.
-        mesh:               The mesh from which the compartments were built.
-        dir_vec:            Direction vector, NxD where N is the number of mesh elements and D the dimension of the mesh.
-        vel_vec:            Velocity vector, NxD where N is the number of mesh elements and D the dimension of the mesh.
-        config_parser:      The OpenCCM ConfigParser
+    Parameters
+    ----------
+    * compartments:         A list of sets. The ith set contains the indices of the elements belonging to the ith compartment.
+    * mesh:                 The mesh from which the compartments were built.
+    * dir_vec:              Direction vector, NxD where N is the number of mesh elements and D the dimension of the mesh.
+    * flows_and_upwind:     2D object array indexed by facet ID.
+                            - 1st column is volumetric flowrate through facet.
+                            - 2nd column is a flag indicating which of a facet's elements are upwind of it.
+                                - 0, and 1 represent the index into mesh.facet_elements[facet]
+                                - -1 is used for boundary elements to represent
+    * config_parser:        The OpenCCM ConfigParser
 
-    Returns:
-        compartments:           Same as input, except renumbered and modified after merging small compartments
-        compartment_network:    A dictionary whose keys are the compartment index and whose values are dictionary.
-                                    For each of those dictionaries, the keys are the index of a neighboring compartment
-                                    and the values are another dictionary.
-                                        For each of those dictionaries, the keys are the index of the bounding facet between the two compartments,
-                                        and the values are Tuples.
-                                            - The 1st is the index of the element upwind of that boundary facet.
-                                            - The 2nd is the outward facing unit normal for that boundary facet.
+    Returns
+    -------
+    * compartments:         Same as input, except renumbered and modified after merging small compartments
+    * compartment_network:  A dictionary whose keys are the compartment index and whose values are dictionary.
+                                For each of those dictionaries, the keys are the index of a neighboring compartment
+                                and the values are another dictionary.
+                                    For each of those dictionaries, the keys are the index of the bounding facet between the two compartments,
+                                    and the values are Tuples.
+                                        - The 1st is the index of the element upwind of that boundary facet.
+                                        - The 2nd is the outward facing unit normal for that boundary facet.
     """
     print("Creating compartment network")
 
@@ -192,15 +202,18 @@ def renumber_compartments(compartments:         Dict[int, Set[int]],
                           compartment_network:  Dict[int, Dict[int, Dict[int, int]]]) \
         -> None:
     """
-    This function re-numbers the compartments from their current numbering, which can include holes for any compartments
+    Re-number compartments from their current numbering, which can include holes for any compartments
     removed or merged, to [0, N-1] inclusive of both ends, where N is the number of compartments.
 
-    Args:
-        compartments:           See documentation in {calculate_compartments}
[…]
