### 2024-08-26 601b7c3
FORMAT: Documentation for config_functions module
