### 2024-08-26 36579f6
FORMAT: Documentation for entry_points.py
