+++ openccm/postprocessing/analysis.py
@@ -14,6 +14,11 @@
 # You should have received a copy of the GNU Lesser General Public License along with OpenCCM. If not, see             #
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
+
+r"""
+Post-processing functions related to performing analysis on the compartmental model and results from simulations on it.
+"""
+
 from collections import defaultdict
 from typing import Dict, List, Optional, Tuple, Set
 
@@ -44,17 +49,19 @@ def network_to_rtd(system_results:   Tuple[
     NOTE: This function assumes the input is a unit step input and is hardcoded into the analysis.
           A smoothed unit step can be used, with only minor impact on the result, as long as it ramps quickly enough.
 
-    Args:
-        system_results: The results from the CM simulation, in the standard format.
-        c_mesh:         The CMesh used to create the compartment network.
-        config_parser:  The OpeCCM ConfigParser used for the simulation and for creating the network.
-        network:        The compartment network on which the results were simulated.
+    Parameters
+    ----------
+    * system_results: The results from the CM simulation, in the standard format.
+    * c_mesh:         The CMesh used to create the compartment network.
+    * config_parser:  The OpeCCM ConfigParser used for the simulation and for creating the network.
+    * network:        The compartment network on which the results were simulated.
 
-    Returns:
-        ~: A single (num species)x(num times)x3 array containing 3 sub-items:
-            [:, :, 0] is time.
-            [:, :, 1] is E(t), the probability density function derived from F(t)
-            [:, :, 2] is F(t), the cumulative probability function, calculated directly.
+    Returns
+    -------
+    * rtd:  A single (num species)x(num times)x3 array containing 3 sub-items:
+            - `[:, :, 0]` is time.
+            - `[:, :, 1]` is E(t), the probability density function derived from F(t)
+            - `[:, :, 2]` is F(t), the cumulative probability function, calculated directly.
     """
     print("Start calculating RTD")
     id_inlet  = c_mesh.grouped_bcs.id(config_parser.get_item(['POST-PROCESSING', 'inlet_bc_name'], str))
@@ -101,11 +108,19 @@ def plot_results(system_results:    Tuple[
                                         Dict[int, List[Tuple[int, int]]]],
                  c_mesh:            CMesh,
                  config_parser:     ConfigParser,
-                 rtd:               Optional[np.ndarray] = None,) -> None:
+                 rtd:               Optional[np.ndarray] = None) -> None:
     """
+    Generate plots of inlet and outlet concentrations in time and optionally for the residence time distributions
+    for each specie.
 
-    Args:
+    NOTE: Requires the optional Matplotlib dependency. Install with `pip install matplotlib`.
 
+    Parameters
+    ----------
+    system_results: The results of running a simulation on the compartmental model.
+    c_mesh:         The CMesh used for creating the compartmental model.
+    config_parser:  The OpenCCM Configparser used for the simulation and for creating the network.
+    rtd:            Optional, the results of running a residence time distribution analysis.
     """
     try:
         import matplotlib.pyplot as plt
@@ -175,19 +190,22 @@ def visualize_model_network(model_network:  Tuple[
                             config_parser:  ConfigParser
                             ) -> None:
     """
-    Take in a network of PFRs/CSTR and create a visual representation of them using the NetworkX package.
+    Take in a network of PFRs/CSTR and create a visual representation of network.
+
+    NOTE: Requires the optional NetworkX dependency. Install with `pip install networkx`.
 
     This representation is really meant for 2D systems since it produces a 2D image.
     If a 3D system is provided the network will be projected on the xy-plane.
 
-    Args:
-        model_network:  The network of reactors (current PFRs/CSTRs) to visualize.
[…]
