+++ openccm/postprocessing/vtu_output.py
@@ -15,6 +15,10 @@
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
 
+r"""
+Functions related to outputting
+"""
+
 import os
 import shutil
 from collections import defaultdict
@@ -49,17 +53,20 @@ internalField   nonuniform List<scalar>
 {}
 (
 """
+"""OpenFOAM header template used to output simulation results for visualization."""
 
 
 def create_element_label_gfu(mesh: 'ngsolve.Mesh') -> 'ngsolve.GridFunction':
     """
     Function which labels each element based on its ID.
 
-    Args:
-        mesh: The Mesh to label
+    Parameters
+    ----------
+    * mesh: The NGSolve Mesh to label.
 
-    Returns:
-        gfu: The GridFunction containing the labeled mesh elements.
+    Returns
+    -------
+    * gfu: The GridFunction containing the labeled mesh elements.
     """
     from ngsolve import L2, GridFunction
 
@@ -79,14 +86,16 @@ def create_compartment_label_gfu(mesh: 'ngsolve.Mesh', compartments: Dict[int, S
 
     Used for visualization.
 
-    Args:
-        mesh:           The mesh to label on
-        compartments:   A list of lists of elements. Each list is a compartment.
+    Parameters
+    ----------
+    * mesh:           The NGSolve mesh used to create the labels.
+    * compartments:   Mapping between compartment ID and the element IDs that belong to that compartment.
 
-    Returns:
-        gfu:            Gridfunction on a 0th order DG L2 space (i.e. 1 value per element) whose
-                        value on each element is the corresponds to the ID of the compartment to which
-                        that element belongs.
+    Returns
+    -------
+    * gfu:  Gridfunction on a 0th order DG L2 space (i.e. 1 value per element) whose
+            value on each element is the corresponds to the ID of the compartment to which
+            that element belongs.
     """
     from ngsolve import L2, GridFunction
 
@@ -108,9 +117,10 @@ def label_elements_openfoam(cmesh: CMesh, config_parser: ConfigParser) -> None:
     """
     Label each mesh element with its ID to help with debugging.
 
-    Args:
-        cmesh:          The CMesh to print out labels for
-        config_parser:  The OpenCCM ConfigParser.
+    Parameters
+    ----------
+    * cmesh:          The CMesh to print out labels for
+    * config_parser:  The OpenCCM ConfigParser.
     """
     # t_span has to be loaded as a float and then converted to str so that zero gets handled correctly.
     # Asking for t_span directly as str results in 0 being returned instead of 0.0, which will cause two zero folders
@@ -151,10 +161,11 @@ def label_compartments_openfoam(output_file_name: str, compartments: Dict[int, S
     Label each mesh element with the ID of the compartment it belongs to.
     A value of -1 is used to represent elements which are not added to any compartment (e.g. those on a no-slip BC).
 
-    Args:
-        output_file_name:   The filename to save the labeled mesh value into.
-        compartments:       The compartments to visualize.
-        config_parser:      The OpenCCM ConfigPararser.
+    Parameters
+    ----------
+    * output_file_name: The filename to save the labeled mesh value into.
+    * compartments:     The compartments to visualize.
+    * config_parser:    The OpenCCM ConfigPararser.
     """
     # t_span has to be loaded as a float and then converted to str so that zero gets handled correctly.
     # Asking for t_span directly as str results in 0 being returned instead of 0.0, which will cause two zero folders
@@ -237,11 +248,12 @@ def cstrs_to_vtu_and_save_openfoam(
     Each mesh cell is labelled based on which compartment it was a part of.
     Cells which were not part of any compartment are labelled with -1.
 
-    Args:
[…]
