### 2024-08-26 acaebc3
FORMAT: Properly format the copyright notices at the top of files
