### 2024-08-26 c46f1df
FORMAT: Add installation instructions, interface documentation, and example documentation
