+++ openccm/docs/simple_reactors_pfr.py
@@ -1,0 +1,30 @@
+########################################################################################################################
+# Copyright 2024 the authors (see AUTHORS file for full list).                                                         #
+#                                                                                                                      #
+#                                                                                                                      #
+# This file is part of OpenCCM.                                                                                        #
+#                                                                                                                      #
+#                                                                                                                      #
+# OpenCCM is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public  #
+# License as published by the Free Software Foundation,either version 2.1 of the License, or (at your option)          #
+# any later version.                                                                                                   #
+#                                                                                                                      #
+# OpenCCM is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied        #
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.                                                     #
+# See the GNU Lesser General Public License for more details.                                                          #
+#                                                                                                                      #
+# You should have received a copy of the GNU Lesser General Public License along with OpenCCM. If not, see             #
+# <https://www.gnu.org/licenses/>.                                                                                     #
+########################################################################################################################
+
+r"""
+This example requires the optional [OpenCMP](https://opencmp.io/index.html) dependency installed,
+see the [Installation Instructions](installation_guide) for help installing it.
+
+To recreate the results of the example below, run the `pfr_analysis.py` script.
+In addition to the images shown in this document, the analysis script can also output a video for each simulation
+showing the concentration throughout the PFR as a function of time.
+To output this video set `save_movie` at the top of the `pfr_analysis.py` script to `True`.
+
+.. include:: ../../examples/simple_reactors/pfr/analysis/pfr_irreversible_example.md
+"""
