+++ openccm/docs/__init__.py
@@ -1,0 +1,31 @@
+########################################################################################################################
+# Copyright 2024 the authors (see AUTHORS file for full list).                                                         #
+#                                                                                                                      #
+#                                                                                                                      #
+# This file is part of OpenCCM.                                                                                        #
+#                                                                                                                      #
+#                                                                                                                      #
+# OpenCCM is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public  #
+# License as published by the Free Software Foundation,either version 2.1 of the License, or (at your option)          #
+# any later version.                                                                                                   #
+#                                                                                                                      #
+# OpenCCM is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied        #
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.                                                     #
+# See the GNU Lesser General Public License for more details.                                                          #
+#                                                                                                                      #
+# You should have received a copy of the GNU Lesser General Public License along with OpenCCM. If not, see             #
+# <https://www.gnu.org/licenses/>.                                                                                     #
+########################################################################################################################
+
+r"""
+Contained within are [installation instructions](/openccm/docs/installation_guide),
+an overview of OpenCCM's [interface](/openccm/docs/package_interface),
+and a detailed description of the different included examples.
+"""
+
+from .installation_guide import *
+from .package_interface import *
+from .pipe_recirc_OpenCMP import *
+from .pipe_recirc_OpenFOAM import *
+from .simple_reactors_cstr import *
+from .simple_reactors_pfr import *
