+++ openccm/system_solvers/pfr_system.py
@@ -15,6 +15,10 @@
 # <https://www.gnu.org/licenses/>.                                                                                     #
 ########################################################################################################################
 
+r"""
+The functions required for solving a simulation on a PFR network.
+"""
+
 from collections import defaultdict
 from typing import Callable, Dict, List, Tuple
 
@@ -43,33 +47,35 @@ def solve_system(
     Function to solve the system. Call with the network representation of the compartment model and it will
     Generate the numerical system to solve, solve it in time, and return the results.
 
-    Args:
-        pfr_network:    The network representation of the PFR as produced by `create_pfr_network`.
-            1. connections:             A dictionary representing the PFR network.
-                                        The keys are the IDs of each PFR, the values are tuples of two dictionaries.
-                                            - The first dictionary is for the inlet(s) of the PFR.
-                                            - The second dictionary is for the outlet(s) of the PFR.
-                                        For both dictionaries, the key is the connection ID
-                                        and the value is the ID of the PFR on the other end of the connection.
-            2. volumes:                 A numpy array of the volume of each PFR indexed by its ID.
-            3. volumetric_flows:        A numpy array of the volumetric flowrate through each connection indexed by its ID.
-            4. compartment_to_pfr_map:  A map between a compartment ID and the PFR IDs of all PFRs in it.
-                                        The PFR IDs are stored in the order in which they appear
-                                        (i.e. the most upstream PFR is first, and the most downstream PFR is last).
-        config_parser:  OpenCCM ConfigParser used for getting settings.
-        grouped_bcs:
+    Parameters
+    ----------
+    * pfr_network:      The network representation of the PFR as produced by `openccm.compartment_models.pfr.create_pfr_network`.
+        1. connections:             A dictionary representing the PFR network.
+                                    The keys are the IDs of each PFR, the values are tuples of two dictionaries.
+                                        - The first dictionary is for the inlet(s) of the PFR.
+                                        - The second dictionary is for the outlet(s) of the PFR.
+                                    For both dictionaries, the key is the connection ID
+                                    and the value is the ID of the PFR on the other end of the connection.
+        2. volumes:                 A numpy array of the volume of each PFR indexed by its ID.
+        3. volumetric_flows:        A numpy array of the volumetric flowrate through each connection indexed by its ID.
+        4. compartment_to_pfr_map:  A map between a compartment ID and the PFR IDs of all PFRs in it.
+                                    The PFR IDs are stored in the order in which they appear
+                                    (i.e. the most upstream PFR is first, and the most downstream PFR is last).
+    * config_parser:    OpenCCM ConfigParser used for getting settings.
+    * grouped_bcs:      GroupedBCs object for identifying which connections belong to domain inlets/outlets.
 
-    Returns:
-        t:          The times at which the ODE was solved (controlled by solve_ivp)
-        c:          The corresponding values for time t.
-        inlet_map:  Map between a domain inlet IDs and the PFRs that they're connected to.
-                        Keys are domain inlet IDs, values are a list of tuples.
-                            The first entry in the tuple is the PFR id,
-                            The second is the id of the connection between the inlet and the PFR.
-        outlet_map: Map between a domain outlet IDs and the PFRs that they're connected to.
-                        Keys are domain inlet IDs, values are a list of tuples.
-                            The first entry in the tuple is the PFR id,
-                            The second is the id of the connection between the outlet and the PFR.
+    Returns
+    -------
[…]

[…]
