### 2024-08-27 uw-comphys/openccm#39 @speth
With a new checkout of commit 51fe6ab, I now get the following error while trying to run `analysis.py` in this example:
```pytb
Using the default value of 5 for SETUP, num_cores.
Using the default value of False for SETUP, DEBUG.
Using the default value of ./ for SETUP, working_directory.
Using the default value of output_ccm/ for SETUP, output_folder_path.
Using the default value of cache/ for SETUP, tmp_folder_path.
Using the default value of log/ for SETUP, log_folder_path.
Using the default value of 1e-15 for COMPARTMENT MODELLING, flow_threshold.
Using the default value of 1e-15 for COMPARTMENT MODELLING, flow_threshold_facet.
Using the default value of 0.05 for COMPARTMENT MODELLING, dist_threshold.
Using the default value of 0.01 for COMPARTMENT MODELLING, atol_opt.
Using the default value of 0.0001 for SIMULATION, first_timestep.
Using the default value of LSODA for SIMULATION, solver.
Using the default value of None for SIMULATION, reactions_file_path.
Using the default value of 1e-06 for SIMULATION, rtol.
Using the default value of 1e-06 for SIMULATION, atol.
Using the default value of all for SIMULATION, t_eval.
Using the default value of True for POST-PROCESSING, save_to_file.
Using the default value of False for POST-PROCESSING, network_diagram.
Using the default value of 0 for POST-PROCESSING, subdivisions.
Using the default value of 1 for POST-PROCESSING, interpolant_order.
Using the default value of compartment_pfr_vtu/ for POST-PROCESSING, vtu_dir.
End COMPARTMENTALIZE
Creating CSTR network
Traceback (most recent call last):
  File "/Users/speth/src/openccm/examples/OpenCMP/pipe_with_recirc_2d/analysis.py", line 202, in <module>
    data_pfr, data_cstr  = rtd_cm()
                           ^^^^^^^^
  File "/Users/speth/src/openccm/examples/OpenCMP/pipe_with_recirc_2d/analysis.py", line 150, in rtd_cm
    run(config_parser)
  File "/Users/speth/src/openccm/openccm/run.py", line 169, in run
    model_network = create_model_network(model, compartments_post, compartment_network, c_mesh, dir_vec, flows_and_upwind, config_parser)
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/speth/src/openccm/openccm/compartment_models/__init__.py", line 43, in create_model_network
    return create_cstr_network(compartments, compartment_network, mesh, flows_and_upwind, dir_vec, config_parser)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/speth/src/openccm/openccm/compartment_models/cstr.py", line 224, in create_cstr_network
    assert len(outlets) > 0
           ^^^^^^^^^^^^^^^^
AssertionError
```
