### 2024-08-27 28854a8
FIX: Update entry points and fix tests to work from entry point.
