+++ pytests/README.md
@@ -1,3 +1,0 @@
-# Instructions
-1. Tests need to be run from within the `pytests/` subdirectory, not from the main directory.
-2. To run the examples, you need to have OpenCMP installed `pip install OpenCMP`.
