### 2024-08-28 uw-comphys/openccm#37 @luohancfd
@Alex-Vasile

I'm struggling with the example cases and need your help to address them.

1. `examples/OpenCMP`: I assume `config_Stokes` generates and initial guess, `config_INS` is for single specie transient incompressible flow simulation, `config_MCINS` is for multi-species one. Although the usage of OpenCMP is beyond the topic of this repo. It's better to document that. I ran the config_Stokes case successfully. But the `config_INS` case took extensive time. Is it really necessary to 5000s? The same reason applies to the `config_MCINS`. I found I can still run the compartment part with the `time_range` reduced to 30s and get similar plot by running `analysis.py`. If it's really necessary, please provide an example output. I found you have result published https://www.sciencedirect.com/science/article/pii/S0098135424000681?via%3Dihub for `config_INS` but nothing for `config_MCINS`. Besides, I'm a little confused why you can compare the compartment modeling with the `config_MCINS`. In the compartment modeling, it seems you are modeling an almost irrevisble reaction  `Na2CO3 + CaCl2 -> 2NaCl   + CaCO3` (I assume this is gas phase reaction). But I couldn't understand what the case  `config_MCINS` is doing. According to this tutorial case https://opencmp.io/tutorials/tutorial_8.html, you are solving an additional mass conservation equation as `a`. But I couldn't find a similar `[SURFACE_RXN]` block. You do have a block named `[FUNCTIONS]/source` in file `model_config`. But it's 0 for `a`.  I personally feel it's OK to not explain every details of a case, especially when the case is targeted for a specific usage. But if you decide to publish your code, it's better to let the user understand what is going on for the example cases.
2. `examples/OpenFOAM`: I built OpenFOAM-12  download link https://openfoam.org/download/12-source/,  sha256 are listed here in case it matters : e59fad54c62e64f1bb89dbaebe5f99a76dc0a6a91d9aad86042a7c4cef6d0744  OpenFOAM-12.tar.gz 8159763451f714387a3846e9c09717533f293107204a3ad470e4acbaa8ce356b  ThirdParty-12.tar.gz. I didn't change the mesh. So I just renamed folder "_0" to "0", and invoke `foamRun` (step 5 in your instruction). I didn't see any difference of `T` and `Vc` for 0 and 300. Is that the expected behavior? In you instruction, you mentioned
> 9. Open `controlDict`, comment out the solve line with `incompressibleFluid` and uncomment the two lines below it defining a new solve and subSulver, finally uncomment the `#includeFunc scalarTransport` line at the bottom of the file.

But the file you posted https://github.com/uw-comphys/openccm/blob/a18be55ddb1fb0f0f501c3f7d38c4b68cab9b51f/examples/OpenFOAM/pipe_with_recirc/system/controlDict#L19-L24

has already uncommented the lines.

If this is a mistake, please fix it. Also, please provide a tagged frozen version for review.  While I'm writting this comment, I see a new PR get merged.

3. `pytests`: I failed to run the tests due to missing files from the `examples` folder. Please correct the examples, or provide independent test files that don't rely on other data/folder.
