### 2024-08-28 336d675
FORMAT: Include config file info in manuscript
