### 2024-08-28 c130002
FORMAT: Clarify instructions on how to run OpenFOAM example.
