### 2024-08-28 uw-comphys/openccm#37 @Alex-Vasile
@luohancfd

1. I looked into it a bit more, and I'll include some of the pre-computed files. To clarify some of your other points.

	- Yes, the Stokes solve is used as an initial condition for the INS solve.
	- No, INS is for only the hydrodynamics with no species transport. MCINS is for n-component reactive flows coupled with INS for hydrodynamics. Here Stokes and INS are used to find the steady state velocity profile in this domain, which one exists. That steady-state velocity profile is then fixed, to speed up the computation, and used by MCINS to calculate the species transport.
	- The time range for Stokes and INS were leftover from when those simulations were transient, the program solves directly for the steady-state flow profile. The value for the time was irrelevant since it wasn't used. I have removed it as it may confuse others. The time it takes to solve the INS simulation is dictated by the size of the resulting matrix and how many non-linear iterations are required to obtain the steady-state.
	- There are two compartmental model configuration files inside `pipe_recird_2d`:
		1. `CONFIG`: Creates a compartmental model and then runs an inert tracer simualtion on it for calculating residence time. This is the same as the CFD simulation described by `CONFIG_MCINS`. This simulation is what gets compared to the CFD results.
		2. `CONFIG_W_RXN`: This also creates a compartmental model but it shows how to run a reactive species simulation, using a reversible reaction and chemical specie variable names. This is simulation is not compared to any CFD.

2. The `0/` directory contains the pre-computed steady-state results. The instructions state that if you wish to use these results and go straight to running the inert tracer CFD simulation to skip to step 13. I have changed the wording to make this more clear.

3. The tests should have all of the data they need, I updated them in another PR to generate any input data they need. However, They need to be run from inside the pytest folder. #43 addresses this issue. Once that's merged the tests can be run using the `openccm-tests` entry point from any directory.

See #49 for changed related to these comments.
