### 2024-09-10 uw-comphys/openccm#43 @speth
> The main directory contains two example config files, `CONFIG` and `CONFIG_REACTIONS`. Would indicating their existence and pointing the reader to it address this comment?

Does the `CONFIG` file in the repository root describe all of the available configuration options? If so, then I would say yes, pointing to these files is sufficient. Repeating this information in the documentation website would also help clarify how to use the software.
