### 2024-09-10 uw-comphys/openccm#43 @speth
Something else I noticed while looking at the reactions config file for the `pipe_with_recirc_2d` example is that it uses species names like `Na2CO3`, while the the "Reaction Configuration File" section of the paper suggests that only letters are allowed in species names, requiring these species names to be replaced with abstract ones like `a` or `b`. Is this a recent change to the code to support a more natural naming convention?
