### 2024-09-10 uw-comphys/openccm#39 @speth
The example is working for me now.
