### 2024-09-11 uw-comphys/openccm#43 @Alex-Vasile
There was a recent code change to improve that, the manuscript was update to reflect it, but a new version has not been generated. It doesn't look like the bot will respond to me in the original issue.
