### 2024-09-17 uw-comphys/openccm#41 @speth
Hi @Alex-Vasile, I am now able to run the test suite successfully. However, I noticed that while the tests run several examples, there are no assertions in any of the tests that would check that the software is generating the expected results. The test suite would pass as long as the example doesn't terminate with an error of some sort.
