### 2024-09-17 uw-comphys/openccm#43 @speth
Thank you, I think all of my original comments have been addressed. I have just two more small items based on looking at the revised manuscript:

- The sample section of the input file (lines 74-84) runs off the right side of the page, so I suppose you need to manually wrap the comments for the sake of the manuscript.
- I'm not sure what is meant by lines 149-150, "Input/Output Concentrations for ‘NaCl’." etc.
