### 2024-09-17 5cfa550
NEW: Add explicit asserts for simple_reactor tests
