+++ openccm/tests/test_run_examples.py
@@ -18,6 +18,9 @@
 ########################################################################################################################
 
 import importlib
+import os
+from typing import Optional
+
 import pytest
 import shutil
 
@@ -41,33 +44,46 @@ def clean_previous_results(configparser: ConfigParser) -> None:
     delete_and_recreate_folder(configparser['SETUP']['output_folder_path'])
 
 
-def clean_and_run(working_directory: str) -> None:
+def clean_and_run(working_directory: str, analysis_script: Optional[str] = None) -> None:
+    os.chdir(Path(__file__).parents[0])
     configparser = ConfigParser(working_directory + 'CONFIG')
     configparser['SETUP']['working_directory'] = working_directory
 
     clean_previous_results(configparser)
-    run(configparser)
+    if analysis_script:
+        analysis_script_path = os.path.abspath(working_directory + analysis_script)
+        spec = importlib.util.spec_from_file_location("module.name", analysis_script_path)
+        module = importlib.util.module_from_spec(spec)
+        spec.loader.exec_module(module)
+        run_analysis = getattr(module, 'run_analysis')
+        run_analysis()
+    else:
+        run(configparser)
 
 
 @pytest.mark.skipif(not importlib.util.find_spec("opencmp"), reason="requires OpenCMP")
 def test_opencmp_cstr_reversible():
-    clean_and_run(rel_path_to_examples + 'simple_reactors/cstr/reversible/')
+    clean_and_run(working_directory=rel_path_to_examples + 'simple_reactors/cstr/reversible/',
+                  analysis_script='cstr_analysis.py')
 
 
 @pytest.mark.skipif(not importlib.util.find_spec("opencmp"), reason="requires OpenCMP")
 def test_opencmp_cstr_irreversible():
-    clean_and_run(rel_path_to_examples + 'simple_reactors/cstr/irreversible/')
+    clean_and_run(working_directory=rel_path_to_examples + 'simple_reactors/cstr/irreversible/',
+                  analysis_script='cstr_analysis.py')
 
 
 @pytest.mark.skipif(not importlib.util.find_spec("opencmp"), reason="requires OpenCMP")
 def test_opencmp_pfr():
-    clean_and_run(rel_path_to_examples + 'simple_reactors/pfr/')
+    clean_and_run(working_directory=rel_path_to_examples + 'simple_reactors/pfr/',
+                  analysis_script='pfr_analysis.py')
 
 
 @pytest.mark.skipif(not importlib.util.find_spec("opencmp"), reason="requires OpenCMP")
 def test_opencmp_recirc():
     from opencmp.run import run as run_opencmp
     from opencmp.config_functions import ConfigParser as OpenCMPConfigParser
+    os.chdir(Path(__file__).parents[0])
 
     path_to_folder = rel_path_to_examples + 'OpenCMP/pipe_with_recirc_2d/'
 
