+++ examples/simple_reactors/pfr/pfr_analysis.py
@@ -27,151 +27,159 @@ from matplotlib import animation
 
 from openccm import run, ConfigParser
 
-# Save movie (transient solution simulation)?
-save_movie = False
 
-# CFD simulation parameters and reaction rate constant
-Vs = 1
-tau = 10
-Q = Vs/tau
-k = 0.1
+def run_analysis(save_movie: bool = False):
+	# CFD simulation parameters and reaction rate constant
+	Vs = 1
+	tau = 10
+	Q = Vs/tau
+	k = 0.1
 
-# Setup analytical solutions
-a_bc = 1 # BC
-a_ic = 0 # IC
-Ca_func = lambda V, t: ( a_ic * np.exp(-k*t) * (1- HV(t-V/Q, 1)) ) + ( a_bc * np.exp(-k*V/Q) * HV(t-V/Q, 1) )
-						
-b_bc = 0 # BC
-b_ic = 0 # IC
-Cb_func = lambda V, t: 2*a_ic * ( 1 - HV(t-V/Q, 1) ) * (1-np.exp(-k*t))  + 2*a_bc*HV(t-V/Q, 1) * (1-np.exp(-k*V/Q)) + (b_bc - b_ic)*HV(t-V/Q, 1) + b_ic
+	# Setup analytical solutions
+	a_bc = 1 # BC
+	a_ic = 0 # IC
+	Ca_func = lambda V, t: ( a_ic * np.exp(-k*t) * (1- HV(t-V/Q, 1)) ) + ( a_bc * np.exp(-k*V/Q) * HV(t-V/Q, 1) )
 
-# Solve for different volume discretization points in PFR
-for pointsPFR in [21, 101, 501]:
-	
-	# Runtime change of points_per_pfr parameter regardless of what configuration file was given.
+	b_bc = 0 # BC
+	b_ic = 0 # IC
+	Cb_func = lambda V, t: 2*a_ic * ( 1 - HV(t-V/Q, 1) ) * (1-np.exp(-k*t))  + 2*a_bc*HV(t-V/Q, 1) * (1-np.exp(-k*V/Q)) + (b_bc - b_ic)*HV(t-V/Q, 1) + b_ic
+
+	old_cwd = os.getcwd()
 	os.chdir(Path(__file__).parents[0])
-	config_parser = ConfigParser('CONFIG')
-	config_parser['SIMULATION']['points_per_pfr'] = str(pointsPFR)
-	run(config_parser)
 
-	# get openccm results for c and t vectors
-	outputs = os.listdir(os.getcwd()+'/output_ccm')
-	t_vec = np.load(os.getcwd()+'/output_ccm/pfr_t.npy')
-	c_vec = np.load(os.getcwd()+'/output_ccm/pfr_concentrations.npy')
+	# Solve for different volume discretization points in PFR
+	for pointsPFR in [21, 101, 501]:
 
-	# get V arrays for function calls
-	V_inlet = np.zeros((len(t_vec), ))
-	V_mid = np.zeros((len(t_vec), )) + 0.5
-	V_outlet = np.ones((len(t_vec), ))
+		# Runtime change of points_per_pfr parameter regardless of what configuration file was given.
+		config_parser = ConfigParser('CONFIG')
+		config_parser['SIMULATION']['points_per_pfr'] = str(pointsPFR)
+		run(config_parser)
 
-	# organize analytical results, use same time domain points as openccm for error calculation purposes
-	Ca_inlet_an		= Ca_func(V_inlet,  t_vec)
-	Ca_mid_an		= Ca_func(V_mid,    t_vec)
[…]
