+++ examples/simple_reactors/cstr/reversible/cstr_analysis.py
@@ -25,87 +25,95 @@ import matplotlib.pyplot as plt
 
 from openccm import run, ConfigParser
 
-# Reaction rate constants. NOTE: rate constants are not synced with reactions config (manual check is needed to ensure they are identical)
-# for large k:
-kf = 1 # s^-1
-kr = 1 # s^-1
-# for small k:
-# kf = 1e-2
-# kr = 1e-4 
 
-# Initial and "boundary" conditions (inlet feed rates)
-tau = 10
-a0, b0 = 0, 0
-aIN, bIN = 1, 0
+def run_analysis():
+	# Reaction rate constants. NOTE: rate constants are not synced with reactions config (manual check is needed to ensure they are identical)
+	# for large k:
+	kf = 1 # s^-1
+	kr = 1 # s^-1
+	# for small k:
+	# kf = 1e-2
+	# kr = 1e-4
 
-# Let C be the total mass, C = A+B.
-c0 = a0 + b0
-cIN = aIN + bIN
+	# Initial and "boundary" conditions (inlet feed rates)
+	tau = 10
+	a0, b0 = 0, 0
+	aIN, bIN = 1, 0
 
-# Useful constants based on hand written derivation
-alpha = -kf - 1/tau 
-alphaSTAR = kr + kf + 1/tau
-omega = a0 - (kr*cIN + aIN/tau)/alphaSTAR - kr*(c0-cIN)/(kr+kf)
+	# Let C be the total mass, C = A+B.
+	c0 = a0 + b0
+	cIN = aIN + bIN
 
-# Define analytical solutions
-A_t = lambda t: omega*np.exp(-alphaSTAR*t) + kr*(c0-cIN)*np.exp(-t/tau)/(kr + kf) + (kr*cIN + aIN/tau)/alphaSTAR
+	# Useful constants based on hand written derivation
+	alpha = -kf - 1/tau
+	alphaSTAR = kr + kf + 1/tau
+	omega = a0 - (kr*cIN + aIN/tau)/alphaSTAR - kr*(c0-cIN)/(kr+kf)
 
-C_t = lambda t: cIN + (c0 - cIN)*np.exp(-t/tau)
+	# Define analytical solutions
+	A_t = lambda t: omega*np.exp(-alphaSTAR*t) + kr*(c0-cIN)*np.exp(-t/tau)/(kr + kf) + (kr*cIN + aIN/tau)/alphaSTAR
 
-B_t = lambda t: C_t(t) - A_t(t)
+	C_t = lambda t: cIN + (c0 - cIN)*np.exp(-t/tau)
 
-# Now get numerical results from openccm
-os.chdir(Path(__file__).parents[0])
-config_parser = ConfigParser('CONFIG')
+	B_t = lambda t: C_t(t) - A_t(t)
 
-for tolerance in [1e-3, 1e-7, 1e-10, 1e-13]:
-	
-	# Runtime change of tolerance parameters regardless of what configuration file was given.
-	tolerance = '{:.0e}'.format(tolerance)
-	config_parser['SIMULATION']['atol'] = tolerance
-	config_parser['SIMULATION']['rtol'] = tolerance
-	run(config_parser)
-	
-	# get openccm results for c and t vectors
-	outputs = os.listdir(os.getcwd()+'/output_ccm')
-	t_vec = np.load(os.getcwd()+'/output_ccm/cstr_t.npy')
-	c_vec = np.load(os.getcwd()+'/output_ccm/cstr_concentrations.npy')
-	c_vec = np.squeeze(c_vec) # remove DOF (just 1 for CSTR)
+	# Now get numerical results from openccm
+	os.chdir(Path(__file__).parents[0])
+	config_parser = ConfigParser('CONFIG')
 
-	# error computation
[…]
