### 2024-09-17 uw-comphys/openccm#41 @Alex-Vasile
Hi @speth, That is partially correct. Yes, there were no asserts written in `test_run_example.py` itself. However, the functions inside the package contain their own asserts to guarantee expected properties of the input/output.

For example, the bottom of `_merge_connections` contains the following checks:
https://github.com/uw-comphys/openccm/blob/d8e29d1976c9b92bb74206c883f064e4fc93aed3/openccm/compartment_models/pfr.py#L616-L645

Similar checks are included across the majority of functions in the package. So, just by running different examples much of the code is tested and each functions' asserts are passed. Having the example run means that many individual tests are passed throughout the package.

To expand on that, I have added, in #52, some explicit asserts to the example files which will now check the output of the simulations.
