+++ .github/workflows/draft-pdf.yml
@@ -15,7 +15,7 @@ jobs:
           journal: joss
           paper-path: publications/2403_JOSS/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: publications/2403_JOSS/paper.pdf
