### 2024-09-18 uw-comphys/openccm#41 @speth
Thanks, I see. Perhaps a bit unconventional as a testing strategy, but I think it's sufficient for the JOSS review requirements. In the long run, I might suggest thinking about also using more explicit unit and regression tests.
