### 2024-09-23 uw-comphys/openccm#37 @luohancfd
@Alex-Vasile

I can run `pytest` successfully but still have issue with the OpenFOAM test. I followed the README.md file and directly run `foamRun`. Here is the result I got for `T` @t=300
![image](https://github.com/user-attachments/assets/48a8424d-6c11-44cd-9903-f9c2a228813a)

When I run the `analysis.py`, I got the following
```
(venv) han:/home/han/Documents/JOSS/openccm/examples/OpenFOAM/pipe_with_recirc/ (main✗) $ python analysis.py
Using the default value of 8 for SETUP, num_cores.
Using the default value of False for SETUP, DEBUG.
Using the default value of ./ for SETUP, working_directory.
Using the default value of output_ccm/ for SETUP, output_folder_path.
Using the default value of cache/ for SETUP, tmp_folder_path.
Using the default value of log/ for SETUP, log_folder_path.
Using the default value of 1e-15 for COMPARTMENT MODELLING, flow_threshold.
Using the default value of 1e-15 for COMPARTMENT MODELLING, flow_threshold_facet.
Using the default value of 0.05 for COMPARTMENT MODELLING, dist_threshold.
Using the default value of 0 for COMPARTMENT MODELLING, rtol_opt.
Using the default value of 0.01 for COMPARTMENT MODELLING, atol_opt.
Using the default value of 0.0001 for SIMULATION, first_timestep.
Using the default value of LSODA for SIMULATION, solver.
Using the default value of None for SIMULATION, reactions_file_path.
Using the default value of 1e-06 for SIMULATION, rtol.
Using the default value of 1e-06 for SIMULATION, atol.
Using the default value of True for POST-PROCESSING, save_to_file.
Using the default value of True for POST-PROCESSING, network_diagram.
Using the default value of 0 for POST-PROCESSING, subdivisions.
Using the default value of 1 for POST-PROCESSING, interpolant_order.
Using the default value of compartment_pfr_vtu/ for POST-PROCESSING, vtu_dir.
Calculating compartments
Traceback (most recent call last):
  File "/home/han/Documents/JOSS/openccm/examples/OpenFOAM/pipe_with_recirc/analysis.py", line 194, in <module>
    data_pfr, data_cstr  = rtd_cm()
                           ^^^^^^^^
  File "/home/han/Documents/JOSS/openccm/examples/OpenFOAM/pipe_with_recirc/analysis.py", line 149, in rtd_cm
    run(config_parser)
  File "/home/han/Documents/JOSS/venv/lib/python3.11/site-packages/openccm/run.py", line 111, in run
    compartments, _ = calculate_compartments(dir_vec, c_mesh, config_parser)
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/han/Documents/JOSS/venv/lib/python3.11/site-packages/openccm/compartmentalize/unidirectional.py", line 286, in calculate_compartments
    compartment_of_removed_elements, valid_elements = _remove_unwanted_elements(mesh, dir_vec)
                                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/han/Documents/JOSS/venv/lib/python3.11/site-packages/openccm/compartmentalize/unidirectional.py", line 968, in _remove_unwanted_elements
    for element_neighbour in mesh.element_connectivity.pop(element):
                             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'tuple' object has no attribute 'pop'
```

Here is a list of pip packages
```
Package            Version
------------------ ---------------
configparser       7.1.0
contourpy          1.3.0
cycler             0.12.1
fonttools          4.53.1
iniconfig          2.0.0
intel-cmplr-lib-ur 2024.2.1
intel-openmp       2024.2.1
kiwisolver         1.4.7
llvmlite           0.43.0
matplotlib         3.9.2
mkl                2024.2.2
mpmath             1.3.0
netgen-mesher      6.2.2404.post16
netgen_occt        7.8.1
ngsolve            6.2.2404.post51
numba              0.60.0
numpy              2.0.2
openccm            0.1
opencmp            1.0.2
packaging          24.1
pillow             10.4.0
pip                24.0
pluggy             1.5.0
pyparsing          3.1.4
pytest             8.3.3
python-dateutil    2.9.0.post0
scipy              1.14.1
setuptools         65.5.0
six                1.16.0
sympy              1.13.3
tbb                2021.13.1
```
