### 2024-10-16 uw-comphys/openccm#37 @Alex-Vasile
@luohancfd Sorry about the late reply.

The CFD results look correct. That semi circular region requires a very long time to fully saturate. The provided example is capped to t=300s to greatly speed up the time taken to perform the analysis. One could extend the simulation time to several thousand seconds, but the analysis results are not substantially changed.

As for the error you're seeing. I believe this is caused by a mixture of  a new version of the analysis script (which you run from `JOSS/openccm/`) and an outdated version of openccm which is called from `JOSS/venv/lib/python3.11`.  The code that is raising the error was changed in #30 (though the error itself appears to be caused by a mix of old code but a new cmesh version).

Please remove the old version from pip, pull the newest version from GitHub, and then pip install. Let me know if the issue persists afterwards.
