### 2024-10-17 uw-comphys/openccm#43 @Alex-Vasile
@speth Has everything in this issue been addressed?
