### 2024-10-20 uw-comphys/openccm#43 @speth
Yes, I believe so.
