### 2024-10-24 0aa8f45
Update paper.md
Related to https://github.com/openjournals/joss-reviews/issues/6963
