### 2024-10-24 a26e37f
Update paper.md
