### 2024-10-24 26fce11
Update setup.cfg
Updated version to 1.0 for JOSS release
