+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = openccm
-version = 0.9
+version = 1.0
 description = OpenCCM is a compartmental modelling software package which is based on recently-developed fully automated flow alignment compartmentalization methods. It is primarily intended for large-scale flow-based processes where there is weak coupling between composition changes, e.g. through (bio)chemical reactions, and convective mass transport in the system. 
 long_description = file: README.md, LICENSE.md
 keywords = reduced-order modelling, compartmentalization, compartment modelling, continuum mechanics
