### 2024-10-24 d8b71ae
Update setup.cfg
corrected license description
