+++ setup.cfg
@@ -4,7 +4,7 @@ version = 1.0
 description = OpenCCM is a compartmental modelling software package which is based on recently-developed fully automated flow alignment compartmentalization methods. It is primarily intended for large-scale flow-based processes where there is weak coupling between composition changes, e.g. through (bio)chemical reactions, and convective mass transport in the system. 
 long_description = file: README.md, LICENSE.md
 keywords = reduced-order modelling, compartmentalization, compartment modelling, continuum mechanics
-license = BSD 3-Clause License
+license = GNU Lesser General Public License v2.1
 classifiers =
     Topic :: Scientific/Engineering :: Physics
     License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)
