### 2024-10-24 0c80903
Update paper.md
additional corrections
