### 2024-10-30 #6963 @kyleniemeyer
Hi @nasserma, I made a few corrections to the paper in https://github.com/uw-comphys/openccm/pull/54, if you could review and merge those.
