### 2024-10-30 #6963 @kyleniemeyer
@openjournals/dev whoops, I missed the backtics around the software title, which are now appearing as `<code>` around the title on the published article page. Can someone fix this, please?
